Extract high-churn routers: spots, solar, satellites, voacap
Move all remaining route handlers out of server.js into dedicated routers:
- spots.js: POTA, DXC/RBN (TCP+SSE), PSK (MQTT+HTTP), WSPR, SOTA, WWFF
- solar.js: solar data, spacewx, SDO frames/images, lunar, dxpeditions,
  contests, propagation, DRAP
- satellites.js: list, positions, passes, TLE, ISS
- voacap.js: SSN, status, prediction matrix

server.js reduced from 4413 to 130 lines — now purely a composition root
(imports, middleware, router mounts, startup, shutdown).
