+++ server.js
@@ -4,19 +4,12 @@ require('dotenv').config();
 const express = require('express');
 const fs = require('fs');
 const path = require('path');
-const { URL } = require('url');
-const https = require('https');
-const { XMLParser } = require('fast-xml-parser');
-const mqtt = require('mqtt');
-const net = require('net');
 const voacap = require('./voacap-bridge.js');
-const satellite = require('satellite.js');
 const { getConfig } = require('./server-config.js');
 const { startListeners } = require('./server-startup.js');
 
 // --- Extracted modules ---
-const { isPrivateIP, resolveHost, secureFetch, fetchJSON, fetchText } = require('./server/services/http-fetch');
-const { registerCache, startEviction, stopEviction } = require('./server/services/cache-store');
+const { startEviction, stopEviction } = require('./server/services/cache-store');
 const cacheControlMiddleware = require('./server/middleware/cache-control');
 const { setupSecurity } = require('./server/middleware/security');
 
@@ -26,6 +19,10 @@ const { setupDocs } = require('./server/routes/meta');
 const callsignRouter = require('./server/routes/callsign');
 const configRouter = require('./server/routes/config');
 const weatherRouter = require('./server/routes/weather');
+const spotsRouter = require('./server/routes/spots');
+const solarRouter = require('./server/routes/solar');
+const satellitesRouter = require('./server/routes/satellites');
+const voacapRouter = require('./server/routes/voacap');
 
 const app = express();
 const config = getConfig();
@@ -33,7 +30,7 @@ const config = getConfig();
 // --- Security, rate limiting, body parsing, static files ---
 setupSecurity(app, config);
 
-// --- Health check (before rate limiter is fine — setupSecurity registers it first) ---
+// --- Health check (before rate limiter) ---
 app.use(metaRouter);
 
 // --- Cache-Control headers ---
@@ -43,3376 +40,15 @@ app.use(cacheControlMiddleware);
 setupDocs(app, __dirname);
 
 // --- Routers ---
+app.use('/api', spotsRouter);
+app.use('/api', solarRouter);
+app.use('/api', satellitesRouter);
 app.use('/api', callsignRouter);
 app.use('/api', configRouter);
+app.use('/api', voacapRouter);
 app.use('/api/weather', weatherRouter);
 
-// Proxy POTA spots API
-app.get('/api/spots', async (req, res) => {
-  try {
-    const data = await fetchJSON('https://api.pota.app/spot/activator');
-    // Normalize frequency from kHz (POTA API) to MHz (consistent with all other sources)
-    const spots = (Array.isArray(data) ? data : []).map(s => ({
-      ...s,
-      frequency: s.frequency ? (parseFloat(s.frequency) / 1000).toFixed(3) : '',
-    }));
-    res.json(spots);
-  } catch (err) {
-    console.error('Error fetching spots:', err.message);
-    res.status(502).json({ error: 'Failed to fetch POTA spots' });
-  }
-});
-
-// --- POTA Spot Submission Proxy ---
-// Posts a spot to api.pota.app on behalf of the user
-app.post('/api/pota/spot', express.json(), async (req, res) => {
-  try {
-    const { activator, spotter, frequency, reference, mode, comments } = req.body;
-
-    // Validate required fields
-    if (!activator || !/^[A-Z0-9]{3,10}$/i.test(activator)) {
-      return res.status(400).json({ error: 'Invalid activator callsign' });
-    }
-    if (!spotter || !/^[A-Z0-9]{3,10}$/i.test(spotter)) {
-      return res.status(400).json({ error: 'Invalid spotter callsign' });
-    }
-    if (!reference || !/^[A-Z0-9]+-\d{4,}$/i.test(reference)) {
-      return res.status(400).json({ error: 'Invalid park reference' });
-    }
-    if (!frequency || isNaN(parseFloat(frequency))) {
-      return res.status(400).json({ error: 'Invalid frequency' });
-    }
-
-    const spotData = {
-      activator: activator.toUpperCase(),
-      spotter: spotter.toUpperCase(),
-      frequency: String(frequency),
-      reference: reference.toUpperCase(),
-      mode: (mode || '').toUpperCase(),
-      comments: (comments || '').slice(0, 60),
-      source: 'HamTab',
-    };
-
-    const resp = await secureFetch('https://api.pota.app/spot', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(spotData),
-    });
-
-    if (resp.ok) {
-      res.json({ success: true });
-    } else {
-      const text = await resp.text().catch(() => '');
-      console.error('[POTA spot] Upstream error:', resp.status, text);
-      res.status(resp.status).json({ error: `POTA API returned ${resp.status}` });
-    }
-  } catch (err) {
-    console.error('[POTA spot] Error:', err.message);
-    res.status(502).json({ error: 'Failed to submit spot' });
-  }
-});
-
-// --- DX Cluster callsign coordinate cache ---
-
-const dxcCallCache = {};  // { 'W1AW': { lat, lon, expires }, ... }
-const DXC_CALL_TTL_OK   = 24 * 60 * 60 * 1000; // 24 hours for successful lookups
-const DXC_CALL_TTL_FAIL =      60 * 60 * 1000; // 1 hour for failed lookups
-
-async function fetchCallCoords(callsign) {
-  const key = callsign.toUpperCase();
-  const cached = dxcCallCache[key];
-  if (cached && Date.now() < cached.expires) return cached;
-  try {
-    let lat = null, lon = null;
-
-    if (isUSCallsign(key)) {
-      // US calls — callook.info (FCC database, richest data)
-      const data = await fetchJSON(`https://callook.info/${encodeURIComponent(key)}/json`);
-      if (data.status === 'VALID' && data.location) {
-        lat = parseFloat(data.location.latitude) || null;
-        lon = parseFloat(data.location.longitude) || null;
-      }
-    } else {
-      // Non-US calls — HamQTH global database
-      const data = await lookupHamqth(key);
-      if (data && data.lat != null && data.lon != null) {
-        lat = data.lat;
-        lon = data.lon;
-      }
-    }
-
-    if (lat !== null && lon !== null) {
-      const entry = { lat, lon, expires: Date.now() + DXC_CALL_TTL_OK };
-      dxcCallCache[key] = entry;
-      return entry;
-    }
-    const entry = { lat: null, lon: null, expires: Date.now() + DXC_CALL_TTL_FAIL };
-    dxcCallCache[key] = entry;
-    return entry;
-  } catch {
-    const entry = { lat: null, lon: null, expires: Date.now() + DXC_CALL_TTL_FAIL };
-    dxcCallCache[key] = entry;
-    return entry;
-  }
-}
-
-// --- DX Cluster spot cache ---
-
-let dxcCache = { data: null, expires: 0 };
-const DXC_CACHE_TTL = 10 * 1000; // 10 seconds — HamQTH minimum polling interval
-
-// --- PSKReporter spot cache ---
-
-let pskCache = { data: null, expires: 0 };
-const PSK_CACHE_TTL = 5 * 60 * 1000; // 5 minutes — PSKReporter recommended minimum
-
-// --- PSKReporter outbound request governor ---
-// Prevents rate-limit blocks when many users share the same container IP.
-// - Split token buckets: separate limits for global feed vs per-callsign heard queries
-// - Circuit breaker: stops upstream calls after consecutive failures, auto-recovers
-// - Request coalescing: duplicate in-flight callsign queries share one upstream call
-// - Stale-while-revalidate: return last known good data when upstream fails
-
-// --- Per-endpoint token buckets ---
-// Split budget so heard queries (many callsigns) don't starve the global feed
-function createTokenBucket(limit, windowMs) {
-  return { limit, windowMs, tokens: limit, resetAt: Date.now() + windowMs };
-}
-
-const pskGlobalBucket  = createTokenBucket(4, 60 * 1000);  // /api/spots/psk — 4 req/60s
-const pskHeardBucket   = createTokenBucket(6, 60 * 1000);  // /api/spots/psk/heard — 6 req/60s
-
-function pskAcquireToken(bucket) {
-  const now = Date.now();
-  if (now >= bucket.resetAt) {
-    bucket.tokens = bucket.limit;
-    bucket.resetAt = now + bucket.windowMs;
-  }
-  if (bucket.tokens > 0) {
-    bucket.tokens--;
-    return true;
-  }
-  return false;
-}
-
-// --- Circuit breaker ---
-// Opens after consecutive upstream failures, serves stale data during cooldown
-const PSK_CB_THRESHOLD = 3;          // consecutive failures before opening
-const PSK_CB_COOLDOWN  = 60 * 1000;  // 60s — half-open attempt interval
-const pskCircuitBreaker = {
-  failures: 0,
-  state: 'closed',       // closed | open | half-open
-  openedAt: 0,
-};
-
-function pskCbAllowRequest() {
-  if (pskCircuitBreaker.state === 'closed') return true;
-  if (pskCircuitBreaker.state === 'open') {
-    // Check if cooldown elapsed → transition to half-open
-    if (Date.now() - pskCircuitBreaker.openedAt >= PSK_CB_COOLDOWN) {
-      pskCircuitBreaker.state = 'half-open';
-      return true; // allow one probe request
-    }
-    return false;
-  }
-  // half-open — allow (one probe already in flight)
-  return true;
-}
-
-function pskCbRecordSuccess() {
-  pskCircuitBreaker.failures = 0;
-  pskCircuitBreaker.state = 'closed';
-}
-
-function pskCbRecordFailure() {
-  pskCircuitBreaker.failures++;
-  if (pskCircuitBreaker.failures >= PSK_CB_THRESHOLD) {
-    pskCircuitBreaker.state = 'open';
-    pskCircuitBreaker.openedAt = Date.now();
-    console.warn(`[PSK] Circuit breaker OPEN after ${pskCircuitBreaker.failures} consecutive failures — serving stale data for ${PSK_CB_COOLDOWN / 1000}s`);
-  }
-}
-
-// Build response metadata for stale/breaker status
-function pskMeta(cache) {
-  if (!cache?.data) return undefined;
-  const ageMs = Date.now() - (cache.expires - PSK_CACHE_TTL);
-  const stale = Date.now() >= cache.expires;
-  if (!stale && pskCircuitBreaker.state === 'closed') return undefined;
-  return { stale, ageSeconds: Math.round(ageMs / 1000), circuitBreaker: pskCircuitBreaker.state };
-}
-
-// In-flight request dedup — keyed by full URL
-const pskInflight = new Map(); // url → Promise<string>
-
-async function pskFetchDedup(url) {
-  if (pskInflight.has(url)) return pskInflight.get(url);
-  const promise = fetchText(url).finally(() => pskInflight.delete(url));
-  pskInflight.set(url, promise);
-  return promise;
-}
-
-// App contact string — PSKReporter requires this for identification
-const PSK_APP_CONTACT = 'appcontact=admin@hamtab.net';
-
-// --- PSKReporter MQTT real-time feed ---
-// Connects to mqtt.pskreporter.info (community service by M0LTE) for real-time spots.
-// Bypasses HTTP retrieval API entirely — no Cloudflare IP blocking issues.
-// Falls back to HTTP retrieval if MQTT is unavailable.
-
-const MQTT_BROKER = 'mqtt://mqtt.pskreporter.info:1883';
-const MQTT_SPOT_TTL = 60 * 60 * 1000; // 1 hour — matches HTTP query window
-const MQTT_RECONNECT_MS = 30 * 1000;  // 30s reconnect delay
-const MQTT_CLEANUP_INTERVAL = 5 * 60 * 1000; // 5 min — purge expired spots
-
-// Per-callsign MQTT spot accumulators
-// { 'CALLSIGN': { spots: Map<key, spot>, subscribed: true, lastAccess: ts } }
-const mqttCallsigns = {};
-
-let mqttClient = null;
-let mqttConnected = false;
-
-// Subscribe to spots where callsign is the SENDER (others hearing you)
-function mqttSubscribeCallsign(callsign) {
-  const upper = callsign.toUpperCase();
-  if (mqttCallsigns[upper]?.subscribed) {
-    mqttCallsigns[upper].lastAccess = Date.now();
-    return;
-  }
-  mqttCallsigns[upper] = {
-    spots: new Map(),
-    subscribed: false,
-    lastAccess: Date.now(),
-  };
-  if (mqttClient && mqttConnected) {
-    const topic = `pskr/filter/v2/+/+/${upper}/#`;
-    mqttClient.subscribe(topic, { qos: 0 }, (err) => {
-      if (!err) {
-        mqttCallsigns[upper].subscribed = true;
-        console.log(`[MQTT] Subscribed: ${topic}`);
-      } else {
-        console.error(`[MQTT] Subscribe error for ${upper}:`, err.message);
-      }
-    });
-  }
-}
-
-// Unsubscribe callsigns not accessed in 30 minutes (cleanup stale subscriptions)
-function mqttCleanupSubscriptions() {
-  const staleThreshold = 30 * 60 * 1000;
-  const now = Date.now();
-  for (const [call, entry] of Object.entries(mqttCallsigns)) {
-    if (now - entry.lastAccess > staleThreshold) {
-      if (mqttClient && mqttConnected && entry.subscribed) {
-        mqttClient.unsubscribe(`pskr/filter/v2/+/+/${call}/#`);
-      }
-      delete mqttCallsigns[call];
-      console.log(`[MQTT] Unsubscribed stale callsign: ${call}`);
-    }
-  }
-}
-
-// Purge spots older than TTL from all accumulators
-function mqttPurgeExpiredSpots() {
-  const cutoff = Date.now() - MQTT_SPOT_TTL;
-  for (const entry of Object.values(mqttCallsigns)) {
-    for (const [key, spot] of entry.spots) {
-      if (spot._receivedAt < cutoff) entry.spots.delete(key);
-    }
-  }
-}
-
-// Convert MQTT JSON message to our internal spot format
-function mqttMessageToSpot(msg) {
-  const freqHz = msg.f || 0;
-  const freqMHz = (freqHz / 1000000).toFixed(3);
-  const spotTime = msg.t ? new Date(msg.t * 1000).toISOString() : new Date().toISOString();
-
-  const senderCoords = gridToLatLon(msg.sl || '');
-  const receiverCoords = gridToLatLon(msg.rl || '');
-
-  const snrVal = typeof msg.rp === 'number' ? msg.rp : parseInt(msg.rp, 10);
-  const snrStr = isNaN(snrVal) ? '' : `${snrVal > 0 ? '+' : ''}${snrVal}`;
-
-  let distanceKm = null;
-  if (senderCoords && receiverCoords) {
-    distanceKm = Math.round(haversineKm(
-      senderCoords.lat, senderCoords.lon,
-      receiverCoords.lat, receiverCoords.lon
-    ));
-  }
-
-  const band = msg.b || freqToBandStr(freqMHz);
-
-  return {
-    callsign: msg.sc || '',
-    frequency: freqMHz,
-    mode: msg.md || '',
-    receiver: msg.rc || '',
-    receiverLocator: msg.rl || '',
-    receiverLat: receiverCoords?.lat || null,
-    receiverLon: receiverCoords?.lon || null,
-    senderLocator: msg.sl || '',
-    senderLat: senderCoords?.lat || null,
-    senderLon: senderCoords?.lon || null,
-    latitude: senderCoords?.lat || null,
-    longitude: senderCoords?.lon || null,
-    snr: snrStr,
-    band,
-    spotTime,
-    distanceKm,
-    reporter: msg.rc || '',
-    reporterLocator: msg.rl || '',
-    reporterLat: receiverCoords?.lat || null,
-    reporterLon: receiverCoords?.lon || null,
-    name: `Heard by ${msg.rc || 'unknown'}`,
-    comments: snrStr ? `SNR: ${snrStr} dB` : '',
-    _receivedAt: Date.now(), // internal timestamp for TTL eviction
-  };
-}
-
-// Build "heard" response from accumulated MQTT spots (same shape as HTTP endpoint)
-function mqttBuildHeardResponse(callsign) {
-  const entry = mqttCallsigns[callsign];
-  if (!entry) return { spots: [], summary: {} };
-
-  const validSpots = [];
-  for (const spot of entry.spots.values()) {
-    if (spot.receiverLat !== null && spot.receiverLon !== null) {
-      validSpots.push(spot);
-    }
-  }
-
-  // Sort newest first
-  validSpots.sort((a, b) => new Date(b.spotTime) - new Date(a.spotTime));
-
-  const summary = {};
-  for (const spot of validSpots) {
-    if (!spot.band) continue;
-    if (!summary[spot.band]) {
-      summary[spot.band] = { count: 0, farthestKm: 0, farthestCall: '' };
-    }
-    summary[spot.band].count++;
-    if (spot.distanceKm !== null && spot.distanceKm > summary[spot.band].farthestKm) {
-      summary[spot.band].farthestKm = spot.distanceKm;
-      summary[spot.band].farthestCall = spot.receiver;
-    }
-  }
-
-  return { spots: validSpots, summary, source: 'mqtt' };
-}
-
-// Initialize MQTT connection
-function initMqttClient() {
-  console.log('[MQTT] Connecting to mqtt.pskreporter.info...');
-  mqttClient = mqtt.connect(MQTT_BROKER, {
-    reconnectPeriod: MQTT_RECONNECT_MS,
-    connectTimeout: 10 * 1000, // 10s connect timeout
-    clientId: `hamtab_${Date.now().toString(36)}`,
-    clean: true,
-  });
-
-  mqttClient.on('connect', () => {
-    mqttConnected = true;
-    console.log('[MQTT] Connected to mqtt.pskreporter.info');
-
-    // Re-subscribe all active callsigns on reconnect
-    for (const [call, entry] of Object.entries(mqttCallsigns)) {
-      const topic = `pskr/filter/v2/+/+/${call}/#`;
-      mqttClient.subscribe(topic, { qos: 0 }, (err) => {
-        if (!err) {
-          entry.subscribed = true;
-        }
-      });
-    }
-
-    // Note: global feed (pskr/filter/v2/#) is NOT subscribed — too much data.
-    // /api/spots/psk uses HTTP fallback. MQTT is for per-callsign Live Spots only.
-  });
-
-  mqttClient.on('message', (topic, payload) => {
-    try {
-      const msg = JSON.parse(payload.toString());
-      const spot = mqttMessageToSpot(msg);
-
-      // Dedupe key: sender-receiver-band-timestamp
-      const key = `${spot.callsign}-${spot.receiver}-${spot.band}-${spot.spotTime}`;
-
-      // Route to per-callsign accumulator if subscribed
-      const senderUpper = (spot.callsign || '').toUpperCase();
-      if (mqttCallsigns[senderUpper]) {
-        mqttCallsigns[senderUpper].spots.set(key, spot);
-        // Cap at 500 spots per callsign
-        if (mqttCallsigns[senderUpper].spots.size > 500) {
-          const oldest = mqttCallsigns[senderUpper].spots.keys().next().value;
-          mqttCallsigns[senderUpper].spots.delete(oldest);
-        }
-      }
-
-    } catch {
-      // Silently skip malformed messages
-    }
-  });
-
-  mqttClient.on('error', (err) => {
-    console.error('[MQTT] Error:', err.message);
-  });
-
-  mqttClient.on('close', () => {
-    mqttConnected = false;
-    // Mark all subscriptions as needing re-subscribe
-    for (const entry of Object.values(mqttCallsigns)) {
-      entry.subscribed = false;
-    }
-  });
-
-  mqttClient.on('offline', () => {
-    mqttConnected = false;
-  });
-
-  // Periodic cleanup
-  setInterval(() => {
-    mqttPurgeExpiredSpots();
-    mqttCleanupSubscriptions();
-  }, MQTT_CLEANUP_INTERVAL);
-}
-
-// Start MQTT on server boot
-initMqttClient();
-
-// --- DX Cluster TCP Telnet + RBN Live Feeds ---
-// Lazy TCP connections to DX Cluster and Reverse Beacon Network.
-// Connected when first SSE client subscribes, disconnected after idle timeout.
-// Uses node:net (no new dependencies). Skipped in hostedmode (no long-lived TCP).
-
-const DXC_TCP_HOST = process.env.DXC_TCP_HOST || 'dxc.ai8w.net';
-const DXC_TCP_PORT = parseInt(process.env.DXC_TCP_PORT, 10) || 7300;
-const RBN_TCP_HOST = 'telnet.reversebeacon.net';
-const RBN_TCP_PORT = 7000;
-const DXC_TCP_RING_SIZE = 500;  // max spots in ring buffer per source
-const DXC_TCP_SPOT_TTL = 60 * 60 * 1000; // 1 hour
-const DXC_TCP_RECONNECT_MS = 30 * 1000;  // 30s backoff
-const DXC_TCP_IDLE_MS = 5 * 60 * 1000;   // 5 min idle disconnect
-const DXC_TCP_HEARTBEAT_MS = 30 * 1000;  // 30s SSE heartbeat
-
-// Shared state for each TCP source
-function createTcpSource() {
-  return {
-    socket: null,
-    connected: false,
-    ring: [],           // ring buffer of parsed spots
-    sseClients: new Set(),
-    reconnectTimer: null,
-    idleTimer: null,
-    lineBuffer: '',     // partial-line accumulator
-    loginSent: false,
-  };
-}
-
-const dxcTcp = createTcpSource();
-const rbnTcp = createTcpSource();
-
-// --- DXC Spot Parser ---
-// Format: "DX de <spotter>:     <freq_khz>  <dx_call>       <comment>           <time>Z"
-// Example: "DX de W3LPL:     14025.0  JA1ABC       CW 10 dB 25 WPM             1423Z"
-const DXC_SPOT_RE = /^DX\s+de\s+(\S+):\s+([\d.]+)\s+(\S+)\s+(.*?)\s+(\d{4})Z\s*$/;
-
-function parseDxcSpot(line) {
-  const m = line.match(DXC_SPOT_RE);
-  if (!m) return null;
-  const [, spotter, freqKhz, dxCall, comment, timeZ] = m;
-  const freqMHz = (parseFloat(freqKhz) / 1000).toFixed(3);
-  const now = new Date();
-  const hh = timeZ.substring(0, 2);
-  const mm = timeZ.substring(2, 4);
-  const spotTime = new Date(Date.UTC(now.getUTCFullYear(), now.getUTCMonth(), now.getUTCDate(),
-    parseInt(hh, 10), parseInt(mm, 10))).toISOString();
-
-  return {
-    callsign: dxCall.trim(),
-    frequency: freqMHz,
-    mode: extractModeFromComment(comment),
-    spotter: spotter.replace(/:$/, ''),
-    band: freqToBandStr(parseFloat(freqMHz)),
-    spotTime,
-    comments: comment.trim(),
-    source: 'dxc-tcp',
-    latitude: null,
-    longitude: null,
-  };
-}
-
-// --- RBN Spot Parser ---
-// Same DX-spot format but often includes dB, WPM, and type in comment
-// Example: "DX de KM3T-2-#:  14040.1  UA3AKO      CW    18 dB  25 WPM  CQ      1423Z"
-const RBN_DB_RE = /(\d+)\s*dB/i;
-const RBN_WPM_RE = /(\d+)\s*WPM/i;
-const RBN_TYPE_RE = /\b(CQ|BEACON|NCDXF|DX)\b/i;
-
-function parseRbnSpot(line) {
-  const base = parseDxcSpot(line);
-  if (!base) return null;
-  base.source = 'rbn';
-
-  // Extract RBN-specific fields from comment
-  const dbMatch = base.comments.match(RBN_DB_RE);
-  const wpmMatch = base.comments.match(RBN_WPM_RE);
-  const typeMatch = base.comments.match(RBN_TYPE_RE);
-
-  base.db = dbMatch ? parseInt(dbMatch[1], 10) : null;
-  base.wpm = wpmMatch ? parseInt(wpmMatch[1], 10) : null;
-  base.spotType = typeMatch ? typeMatch[1].toUpperCase() : '';
-  base.skimmer = base.spotter; // RBN spotters are CW Skimmer stations
-
-  return base;
-}
-
-// --- Ring Buffer Management ---
-
-function ringPush(source, spot) {
-  spot._ts = Date.now();
-  source.ring.push(spot);
-  // Enforce size limit
-  if (source.ring.length > DXC_TCP_RING_SIZE) {
-    source.ring.shift();
-  }
-  // Purge expired spots
-  const cutoff = Date.now() - DXC_TCP_SPOT_TTL;
-  while (source.ring.length > 0 && source.ring[0]._ts < cutoff) {
-    source.ring.shift();
-  }
-}
-
-// --- TCP Connection Manager ---
-
-function connectTcpSource(source, host, port, callsign, parseFn) {
-  if (process.env.HOSTED_MODE === '1') return; // no TCP in hostedmode
-  if (source.socket) return; // already connected or connecting
-  if (source.reconnectTimer) { clearTimeout(source.reconnectTimer); source.reconnectTimer = null; }
-
-  const loginCall = callsign || process.env.DXC_CALLSIGN || process.env.HAMTAB_CALLSIGN || 'HAMTAB';
-  console.log(`[TCP] Connecting to ${host}:${port} as ${loginCall}`);
-
-  const sock = net.createConnection({ host, port, timeout: 15000 });
-  source.socket = sock;
-  source.loginSent = false;
-  source.lineBuffer = '';
-
-  sock.setEncoding('utf8');
-
-  sock.on('connect', () => {
-    source.connected = true;
-    console.log(`[TCP] Connected to ${host}:${port}`);
-    broadcastSseStatus(source, { connected: true, host });
-  });
-
-  sock.on('data', (chunk) => {
-    source.lineBuffer += chunk;
-    const lines = source.lineBuffer.split('\n');
-    source.lineBuffer = lines.pop(); // keep incomplete last line
-
-    for (const rawLine of lines) {
-      const line = rawLine.replace(/\r/g, '').trim();
-      if (!line) continue;
-
-      // Login prompt detection — send callsign once
-      if (!source.loginSent && (line.includes('login:') || line.includes('call:') || line.includes('Please enter your call'))) {
-        sock.write(loginCall + '\r\n');
-        source.loginSent = true;
-        continue;
-      }
-
-      // Parse spot
-      const spot = parseFn(line);
-      if (spot) {
-        ringPush(source, spot);
-        broadcastSseSpot(source, spot);
-      }
-    }
-  });
-
-  sock.on('timeout', () => {
-    console.warn(`[TCP] Connection to ${host}:${port} timed out`);
-    sock.destroy();
-  });
-
-  sock.on('error', (err) => {
-    console.error(`[TCP] Error on ${host}:${port}:`, err.message);
-  });
-
-  sock.on('close', () => {
-    source.connected = false;
-    source.socket = null;
-    source.loginSent = false;
-    source.lineBuffer = '';
-    console.log(`[TCP] Disconnected from ${host}:${port}`);
-    broadcastSseStatus(source, { connected: false, host });
-
-    // Reconnect if there are still SSE clients
-    if (source.sseClients.size > 0) {
-      source.reconnectTimer = setTimeout(() => {
-        connectTcpSource(source, host, port, callsign, parseFn);
-      }, DXC_TCP_RECONNECT_MS);
-    }
-  });
-
-  // Reset idle timer
-  resetIdleTimer(source, host, port);
-}
-
-function disconnectTcpSource(source) {
-  if (source.reconnectTimer) { clearTimeout(source.reconnectTimer); source.reconnectTimer = null; }
-  if (source.idleTimer) { clearTimeout(source.idleTimer); source.idleTimer = null; }
-  if (source.socket) {
-    source.socket.destroy();
-    source.socket = null;
-  }
-  source.connected = false;
-  source.loginSent = false;
-  source.lineBuffer = '';
-}
-
-function resetIdleTimer(source, host, port) {
-  if (source.idleTimer) clearTimeout(source.idleTimer);
-  source.idleTimer = setTimeout(() => {
-    if (source.sseClients.size === 0) {
-      console.log(`[TCP] Idle timeout — disconnecting from ${host}:${port}`);
-      disconnectTcpSource(source);
-    }
-  }, DXC_TCP_IDLE_MS);
-}
-
-// --- SSE Broadcasting ---
-
-function broadcastSseSpot(source, spot) {
-  const data = JSON.stringify(spot);
-  for (const res of source.sseClients) {
-    res.write(`data: ${data}\n\n`);
-  }
-}
-
-function broadcastSseStatus(source, status) {
-  const data = JSON.stringify(status);
-  for (const res of source.sseClients) {
-    res.write(`event: status\ndata: ${data}\n\n`);
-  }
-}
-
-function addSseClient(source, res, host, port, callsign, parseFn) {
-  // SSE headers
-  res.writeHead(200, {
-    'Content-Type': 'text/event-stream',
-    'Cache-Control': 'no-cache',
-    'Connection': 'keep-alive',
-    'X-Accel-Buffering': 'no', // nginx proxy compat
-  });
-
-  // Send current ring buffer as init event
-  const initData = JSON.stringify(source.ring.map(s => { const { _ts, ...rest } = s; return rest; }));
-  res.write(`event: init\ndata: ${initData}\n\n`);
-
-  // Send current connection status
-  res.write(`event: status\ndata: ${JSON.stringify({ connected: source.connected, host })}\n\n`);
-
-  source.sseClients.add(res);
-
-  // Heartbeat to keep connection alive
-  const heartbeat = setInterval(() => {
-    res.write(`event: heartbeat\ndata: \n\n`);
-  }, DXC_TCP_HEARTBEAT_MS);
-
-  // Start TCP connection if not already connected
-  if (!source.socket) {
-    connectTcpSource(source, host, port, callsign, parseFn);
-  }
-
-  // Cleanup on client disconnect
-  res.on('close', () => {
-    source.sseClients.delete(res);
-    clearInterval(heartbeat);
-    // Start idle timer if no clients remain
-    if (source.sseClients.size === 0) {
-      resetIdleTimer(source, host, port);
-    }
-  });
-}
-
-// --- SSE Endpoints ---
-
-if (process.env.HOSTED_MODE !== '1') {
-  // DXC live TCP stream
-  app.get('/api/spots/dxc/stream', (req, res) => {
-    const callsign = (req.query.callsign || '').replace(/[^A-Z0-9/]/gi, '').substring(0, 10);
-    addSseClient(dxcTcp, res, DXC_TCP_HOST, DXC_TCP_PORT, callsign, parseDxcSpot);
-  });
-
-  // RBN live TCP stream
-  app.get('/api/spots/rbn/stream', (req, res) => {
-    const callsign = (req.query.callsign || '').replace(/[^A-Z0-9/]/gi, '').substring(0, 10);
-    addSseClient(rbnTcp, res, RBN_TCP_HOST, RBN_TCP_PORT, callsign, parseRbnSpot);
-  });
-
-  // RBN HTTP snapshot (polling fallback)
-  app.get('/api/spots/rbn', (req, res) => {
-    res.json(rbnTcp.ring.map(s => { const { _ts, ...rest } = s; return rest; }));
-  });
-}
-
-// Extract mode from DXC comment field
-function extractModeFromComment(comment) {
-  if (!comment) return '';
-  const c = comment.toUpperCase();
-  // Common digital modes
-  if (c.includes('FT8')) return 'FT8';
-  if (c.includes('FT4')) return 'FT4';
-  if (c.includes('JS8')) return 'JS8';
-  if (c.includes('RTTY')) return 'RTTY';
-  if (c.includes('PSK31') || c.includes('PSK')) return 'PSK';
-  if (c.includes('JT65')) return 'JT65';
-  if (c.includes('JT9')) return 'JT9';
-  if (c.includes('WSPR')) return 'WSPR';
-  if (c.includes('MFSK')) return 'MFSK';
-  if (c.includes('OLIVIA')) return 'OLIVIA';
-  // CW
-  if (c.includes('CW') || c.includes('MORSE')) return 'CW';
-  // Phone modes
-  if (c.includes('SSB') || c.includes('USB') || c.includes('LSB')) return 'SSB';
-  if (c.includes('FM')) return 'FM';
-  if (c.includes('AM')) return 'AM';
-  return '';
-}
-
-// Convert frequency to band string
-function freqToBandStr(freqMHz) {
-  const freq = parseFloat(freqMHz);
-  if (isNaN(freq)) return '';
-  if (freq >= 1.8 && freq <= 2.0) return '160m';
-  if (freq >= 3.5 && freq <= 4.0) return '80m';
-  if (freq >= 5.3 && freq <= 5.4) return '60m';
-  if (freq >= 7.0 && freq <= 7.3) return '40m';
-  if (freq >= 10.1 && freq <= 10.15) return '30m';
-  if (freq >= 14.0 && freq <= 14.35) return '20m';
-  if (freq >= 18.068 && freq <= 18.168) return '17m';
-  if (freq >= 21.0 && freq <= 21.45) return '15m';
-  if (freq >= 24.89 && freq <= 24.99) return '12m';
-  if (freq >= 28.0 && freq <= 29.7) return '10m';
-  if (freq >= 50.0 && freq <= 54.0) return '6m';
-  if (freq >= 144.0 && freq <= 148.0) return '2m';
-  if (freq >= 420.0 && freq <= 450.0) return '70cm';
-  return '';
-}
-
-// (Callsign helpers + HamQTH lookup moved to server/routes/callsign.js)
-
-// Proxy HamQTH DX Cluster spots
-app.get('/api/spots/dxc', async (req, res) => {
-  try {
-    // Return cached data if fresh
-    if (dxcCache.data && Date.now() < dxcCache.expires) {
-      return res.json(dxcCache.data);
-    }
-
-    const raw = await fetchText('https://www.hamqth.com/dxc_csv.php');
-    const lines = raw.split('\n').filter(l => l.trim());
-
-    const spots = [];
-    for (const line of lines) {
-      // HamQTH CSV is caret-delimited: callsign^freq^datetime^spotter^comment^lotw^eqsl^continent^band^country^adif_id
-      const parts = line.split('^');
-      if (parts.length < 11) continue;
-
-      const [callsign, freqKhz, datetime, spotter, comment, lotw, eqsl, continent, band, country, adifId] = parts;
-
-      // Convert frequency from kHz to MHz
-      const freqMHz = (parseFloat(freqKhz) / 1000).toFixed(3);
-
-      // Parse datetime (format: YYYY-MM-DD HH:MM:SS)
-      let spotTime = '';
-      if (datetime) {
-        spotTime = datetime.replace(' ', 'T') + 'Z';
-      }
-
-      spots.push({
-        callsign: callsign || '',
-        frequency: freqMHz,
-        mode: extractModeFromComment(comment),
-        spotter: spotter || '',
-        name: country || '',
-        continent: continent || '',
-        band: freqToBandStr(parseFloat(freqMHz)) || band || '',
-        spotTime,
-        comments: comment || '',
-        lotwUser: lotw === '1',
-        eqslUser: eqsl === '1',
-        latitude: null,
-        longitude: null,
-        adifId: adifId || '',
-      });
-    }
-
-    // Batch lookup coordinates for unique US callsigns
-    const usCallsigns = new Set();
-    for (const s of spots) {
-      if (s.callsign && /^[AKNW][A-Z]?\d/.test(s.callsign.toUpperCase())) {
-        usCallsigns.add(s.callsign.toUpperCase());
-      }
-    }
-
-    // Limit concurrent lookups to avoid overwhelming callook.info
-    const callsignArray = [...usCallsigns].slice(0, 50); // max 50 lookups per refresh
-    await Promise.allSettled(callsignArray.map(c => fetchCallCoords(c)));
-
-    // Merge coordinates
-    for (const s of spots) {
-      const key = s.callsign.toUpperCase();
-      const coords = dxcCallCache[key];
-      if (coords && coords.lat !== null && coords.lon !== null) {
-        s.latitude = coords.lat;
-        s.longitude = coords.lon;
-      }
-    }
-
-    // Cache the result
-    dxcCache = { data: spots, expires: Date.now() + DXC_CACHE_TTL };
-
-    res.json(spots);
-  } catch (err) {
-    console.error('Error fetching DXC spots:', err.message);
-    res.status(502).json({ error: 'Failed to fetch DX Cluster spots' });
-  }
-});
-
-// Proxy PSKReporter API
-app.get('/api/spots/psk', async (req, res) => {
-  try {
-    // Return cached data if fresh
-    if (pskCache.data && Date.now() < pskCache.expires) {
-      return res.json(pskCache.data);
-    }
-
-    // Circuit breaker check — serve stale if breaker is open
-    // Global PSK returns a plain array (consumed by source tab), so no _meta wrapper
-    if (!pskCbAllowRequest()) {
-      if (pskCache.data) return res.json(pskCache.data);
-      return res.status(503).json({ error: 'PSKReporter circuit breaker open — try again shortly' });
-    }
-
-    // Rate limit check — serve stale data if throttled
-    if (!pskAcquireToken(pskGlobalBucket)) {
-      if (pskCache.data) return res.json(pskCache.data); // stale-while-revalidate
-      return res.status(429).json({ error: 'PSKReporter rate limit — try again shortly' });
-    }
-
-    // Build query params
-    const params = new URLSearchParams({
-      flowStartSeconds: '-3600',  // Last hour
-      rrlimit: '500',             // Max 500 reports
-      rronly: '1',                // Reception reports only
-      nolocator: '0',             // Include grid squares
-    });
-
-    const url = `https://retrieve.pskreporter.info/query?${params}&${PSK_APP_CONTACT}`;
-    const xml = await pskFetchDedup(url);
-
-    pskCbRecordSuccess(); // upstream responded
-
-    // Parse XML
-    const parser = new XMLParser({
-      ignoreAttributes: false,
-      attributeNamePrefix: '',
-    });
-    const doc = parser.parse(xml);
-
-    // Extract reception reports array
-    const reports = doc.receptionReports?.receptionReport;
-    if (!reports) {
-      pskCache = { data: [], expires: Date.now() + PSK_CACHE_TTL };
-      return res.json([]);
-    }
-
-    const reportArray = Array.isArray(reports) ? reports : [reports];
-
-    // Transform to spot format
-    const spots = reportArray.map(r => {
-      const freqHz = parseInt(r.frequency, 10) || 0;
-      const freqMHz = (freqHz / 1000000).toFixed(3);
-      const flowStartSec = parseInt(r.flowStartSeconds, 10) || 0;
-      const spotTime = new Date(flowStartSec * 1000).toISOString();
-
-      const senderCoords = gridToLatLon(r.senderLocator || '');
-      const receiverCoords = gridToLatLon(r.receiverLocator || '');
-
-      const snrVal = parseInt(r.sNR, 10);
-      const snrStr = isNaN(snrVal) ? '' : `${snrVal > 0 ? '+' : ''}${snrVal}`;
-
-      return {
-        callsign: r.senderCallsign || '',
-        frequency: freqMHz,
-        mode: r.mode || '',
-        reporter: r.receiverCallsign || '',
-        reporterLocator: r.receiverLocator || '',
-        reporterLat: receiverCoords?.lat || null,
-        reporterLon: receiverCoords?.lon || null,
-        senderLocator: r.senderLocator || '',
-        latitude: senderCoords?.lat || null,
-        longitude: senderCoords?.lon || null,
-        snr: snrStr,
-        band: freqToBandStr(freqMHz),
-        spotTime,
-        name: `Heard by ${r.receiverCallsign || 'unknown'}`,
-        comments: snrStr ? `SNR: ${snrStr} dB` : '',
-      };
-    });
-
-    // Filter out spots with no valid transmitter coordinates
-    const validSpots = spots.filter(s => s.latitude !== null && s.longitude !== null);
-
-    pskCache = { data: validSpots, expires: Date.now() + PSK_CACHE_TTL };
-    res.json(validSpots);
-  } catch (err) {
-    console.error('Error fetching PSKReporter spots:', err.message);
-    pskCbRecordFailure(); // track consecutive failures
-    // Stale-while-revalidate — serve old data if available
-    if (pskCache.data) return res.json(pskCache.data);
-    res.status(502).json({ error: 'PSKReporter is unavailable — try again shortly' });
-  }
-});
-
-// --- Live Spots (PSKReporter "heard" query) ---
-
-// Per-callsign cache for Live Spots
-const pskHeardCache = {}; // { 'CALLSIGN': { data, expires }, ... }
-const PSK_HEARD_CACHE_TTL = 5 * 60 * 1000; // 5 minutes
-
-// Haversine distance calculation (km)
-function haversineKm(lat1, lon1, lat2, lon2) {
-  const R = 6371; // Earth radius in km
-  const dLat = (lat2 - lat1) * Math.PI / 180;
-  const dLon = (lon2 - lon1) * Math.PI / 180;
-  const a = Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-    Math.cos(lat1 * Math.PI / 180) * Math.cos(lat2 * Math.PI / 180) *
-    Math.sin(dLon / 2) * Math.sin(dLon / 2);
-  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-  return R * c;
-}
-
-// Live Spots: Query PSKReporter for spots where YOUR signal was received
-app.get('/api/spots/psk/heard', async (req, res) => {
-  const callsign = (req.query.callsign || '').toUpperCase().trim();
-  try {
-
-    // Validate callsign format
-    if (!callsign || !/^[A-Z0-9]{3,10}$/i.test(callsign)) {
-      return res.status(400).json({ error: 'Provide a valid callsign' });
-    }
-
-    // Ensure MQTT subscription for this callsign (idempotent)
-    mqttSubscribeCallsign(callsign);
-
-    // Prefer MQTT data if connected and has spots for this callsign
-    if (mqttConnected && mqttCallsigns[callsign]?.spots.size > 0) {
-      const result = mqttBuildHeardResponse(callsign);
-      // Also update the HTTP cache so stale fallback has fresh data
-      pskHeardCache[callsign] = { data: result, expires: Date.now() + PSK_HEARD_CACHE_TTL };
-      return res.json(result);
-    }
-
-    // Check cache
-    const cached = pskHeardCache[callsign];
-    if (cached && Date.now() < cached.expires) {
-      return res.json(cached.data);
-    }
-
-    // Circuit breaker check — serve stale if breaker is open
-    if (!pskCbAllowRequest()) {
-      const meta = pskMeta(cached);
-      if (cached?.data) return res.json({ ...cached.data, _meta: meta });
-      return res.status(503).json({ error: 'PSKReporter circuit breaker open — try again shortly' });
-    }
-
-    // Rate limit check — serve stale data if throttled
-    if (!pskAcquireToken(pskHeardBucket)) {
-      if (cached?.data) return res.json(cached.data); // stale-while-revalidate
-      return res.status(429).json({ error: 'PSKReporter rate limit — try again shortly' });
-    }
-
-    // Build query params for PSKReporter
-    const params = new URLSearchParams({
-      senderCallsign: callsign,
-      flowStartSeconds: '-3600', // Last hour
-      rrlimit: '500',
-      rronly: '1',
-      nolocator: '0',
-    });
-
-    const url = `https://retrieve.pskreporter.info/query?${params}&${PSK_APP_CONTACT}`;
-    const xml = await pskFetchDedup(url);
-
-    pskCbRecordSuccess(); // upstream responded
-
-    // Parse XML
-    const parser = new XMLParser({
-      ignoreAttributes: false,
-      attributeNamePrefix: '',
-    });
-    const doc = parser.parse(xml);
-
-    const reports = doc.receptionReports?.receptionReport;
-    if (!reports) {
-      const result = { spots: [], summary: {} };
-      pskHeardCache[callsign] = { data: result, expires: Date.now() + PSK_HEARD_CACHE_TTL };
-      return res.json(result);
-    }
-
-    const reportArray = Array.isArray(reports) ? reports : [reports];
-
-    // Transform to spot format
-    const spots = reportArray.map(r => {
-      const freqHz = parseInt(r.frequency, 10) || 0;
-      const freqMHz = (freqHz / 1000000).toFixed(3);
-      const flowStartSec = parseInt(r.flowStartSeconds, 10) || 0;
-      const spotTime = new Date(flowStartSec * 1000).toISOString();
-
-      const senderCoords = gridToLatLon(r.senderLocator || '');
-      const receiverCoords = gridToLatLon(r.receiverLocator || '');
-
-      const snrVal = parseInt(r.sNR, 10);
-      const snrStr = isNaN(snrVal) ? '' : `${snrVal > 0 ? '+' : ''}${snrVal}`;
-
-      // Calculate distance if both coords available
-      let distanceKm = null;
-      if (senderCoords && receiverCoords) {
-        distanceKm = Math.round(haversineKm(
-          senderCoords.lat, senderCoords.lon,
-          receiverCoords.lat, receiverCoords.lon
-        ));
-      }
-
-      return {
-        callsign: r.senderCallsign || '',
-        frequency: freqMHz,
-        mode: r.mode || '',
-        receiver: r.receiverCallsign || '',
-        receiverLocator: r.receiverLocator || '',
-        receiverLat: receiverCoords?.lat || null,
-        receiverLon: receiverCoords?.lon || null,
-        senderLocator: r.senderLocator || '',
-        senderLat: senderCoords?.lat || null,
-        senderLon: senderCoords?.lon || null,
-        snr: snrStr,
-        band: freqToBandStr(freqMHz),
-        spotTime,
-        distanceKm,
-      };
-    });
-
-    // Filter out spots with no valid receiver coordinates
-    const validSpots = spots.filter(s => s.receiverLat !== null && s.receiverLon !== null);
-
-    // Build summary by band
-    const summary = {};
-    for (const spot of validSpots) {
-      if (!spot.band) continue;
-      if (!summary[spot.band]) {
-        summary[spot.band] = { count: 0, farthestKm: 0, farthestCall: '' };
-      }
-      summary[spot.band].count++;
-      if (spot.distanceKm !== null && spot.distanceKm > summary[spot.band].farthestKm) {
-        summary[spot.band].farthestKm = spot.distanceKm;
-        summary[spot.band].farthestCall = spot.receiver;
-      }
-    }
-
-    const result = { spots: validSpots, summary };
-    pskHeardCache[callsign] = { data: result, expires: Date.now() + PSK_HEARD_CACHE_TTL };
-    res.json(result);
-  } catch (err) {
-    console.error('Error fetching PSKReporter heard spots:', err.message);
-    pskCbRecordFailure(); // track consecutive failures
-    // Stale-while-revalidate — serve old data if available
-    const stale = pskHeardCache[callsign];
-    if (stale?.data) return res.json(stale.data);
-    res.status(502).json({ error: 'PSKReporter is unavailable — try again shortly' });
-  }
-});
-
-// --- Internal PSKReporter XML parse endpoints (hostedmode Worker edge-fetch) ---
-// The Worker fetches PSKReporter from edge IPs (avoids container IP blocking),
-// then POSTs the raw XML here for parsing. Not called in lanmode.
-
-// Shared parser: PSKReporter XML → spot array
-function parsePskSpots(xml) {
-  const parser = new XMLParser({ ignoreAttributes: false, attributeNamePrefix: '' });
-  const doc = parser.parse(xml);
-  const reports = doc.receptionReports?.receptionReport;
-  if (!reports) return [];
-  return Array.isArray(reports) ? reports : [reports];
-}
-
-// Shared transformer: raw report → API spot format (for /api/spots/psk)
-function transformPskSpot(r) {
-  const freqHz = parseInt(r.frequency, 10) || 0;
-  const freqMHz = (freqHz / 1000000).toFixed(3);
-  const flowStartSec = parseInt(r.flowStartSeconds, 10) || 0;
-  const spotTime = new Date(flowStartSec * 1000).toISOString();
-  const senderCoords = gridToLatLon(r.senderLocator || '');
-  const receiverCoords = gridToLatLon(r.receiverLocator || '');
-  const snrVal = parseInt(r.sNR, 10);
-  const snrStr = isNaN(snrVal) ? '' : `${snrVal > 0 ? '+' : ''}${snrVal}`;
-  return {
-    callsign: r.senderCallsign || '',
-    frequency: freqMHz,
-    mode: r.mode || '',
-    reporter: r.receiverCallsign || '',
-    reporterLocator: r.receiverLocator || '',
-    reporterLat: receiverCoords?.lat || null,
-    reporterLon: receiverCoords?.lon || null,
-    senderLocator: r.senderLocator || '',
-    latitude: senderCoords?.lat || null,
-    longitude: senderCoords?.lon || null,
-    snr: snrStr,
-    band: freqToBandStr(freqMHz),
-    spotTime,
-    name: `Heard by ${r.receiverCallsign || 'unknown'}`,
-    comments: snrStr ? `SNR: ${snrStr} dB` : '',
-  };
-}
-
-// Shared transformer: raw report → heard spot format (for /api/spots/psk/heard)
-function transformPskHeardSpot(r) {
-  const freqHz = parseInt(r.frequency, 10) || 0;
-  const freqMHz = (freqHz / 1000000).toFixed(3);
-  const flowStartSec = parseInt(r.flowStartSeconds, 10) || 0;
-  const spotTime = new Date(flowStartSec * 1000).toISOString();
-  const senderCoords = gridToLatLon(r.senderLocator || '');
-  const receiverCoords = gridToLatLon(r.receiverLocator || '');
-  const snrVal = parseInt(r.sNR, 10);
-  const snrStr = isNaN(snrVal) ? '' : `${snrVal > 0 ? '+' : ''}${snrVal}`;
-  let distanceKm = null;
-  if (senderCoords && receiverCoords) {
-    distanceKm = Math.round(haversineKm(
-      senderCoords.lat, senderCoords.lon,
-      receiverCoords.lat, receiverCoords.lon
-    ));
-  }
-  return {
-    callsign: r.senderCallsign || '',
-    frequency: freqMHz,
-    mode: r.mode || '',
-    receiver: r.receiverCallsign || '',
-    receiverLocator: r.receiverLocator || '',
-    receiverLat: receiverCoords?.lat || null,
-    receiverLon: receiverCoords?.lon || null,
-    senderLocator: r.senderLocator || '',
-    senderLat: senderCoords?.lat || null,
-    senderLon: senderCoords?.lon || null,
-    snr: snrStr,
-    band: freqToBandStr(freqMHz),
-    spotTime,
-    distanceKm,
-  };
-}
-
-// Internal: parse PSK spots XML (Worker edge-fetch → container parse)
-app.post('/api/internal/psk-parse', express.text({ type: '*/*', limit: '5mb' }), (req, res) => {
-  try {
-    const reports = parsePskSpots(req.body);
-    const spots = reports.map(transformPskSpot);
-    const validSpots = spots.filter(s => s.latitude !== null && s.longitude !== null);
-    pskCache = { data: validSpots, expires: Date.now() + PSK_CACHE_TTL };
-    res.json(validSpots);
-  } catch (err) {
-    console.error('[internal/psk-parse] Error:', err.message);
-    res.status(500).json({ error: 'XML parse failed' });
-  }
-});
-
-// Internal: parse PSK heard spots XML (Worker edge-fetch → container parse)
-app.post('/api/internal/psk-heard-parse', express.text({ type: '*/*', limit: '5mb' }), (req, res) => {
-  try {
-    const callsign = (req.query.callsign || '').toUpperCase().trim();
-    const reports = parsePskSpots(req.body);
-    const spots = reports.map(transformPskHeardSpot);
-    const validSpots = spots.filter(s => s.receiverLat !== null && s.receiverLon !== null);
-    const summary = {};
-    for (const spot of validSpots) {
-      if (!spot.band) continue;
-      if (!summary[spot.band]) {
-        summary[spot.band] = { count: 0, farthestKm: 0, farthestCall: '' };
-      }
-      summary[spot.band].count++;
-      if (spot.distanceKm !== null && spot.distanceKm > summary[spot.band].farthestKm) {
-        summary[spot.band].farthestKm = spot.distanceKm;
-        summary[spot.band].farthestCall = spot.receiver;
-      }
-    }
-    const result = { spots: validSpots, summary };
-    if (callsign) pskHeardCache[callsign] = { data: result, expires: Date.now() + PSK_HEARD_CACHE_TTL };
-    res.json(result);
-  } catch (err) {
-    console.error('[internal/psk-heard-parse] Error:', err.message);
-    res.status(500).json({ error: 'XML parse failed' });
-  }
-});
-
-// --- WSPR Spots (wspr.live ClickHouse) ---
-let wsprCache = { data: null, expires: 0 };
-const WSPR_CACHE_TTL = 5 * 60 * 1000; // 5 min — WSPR 2-min cycle; 20 req/min API limit
-
-app.get('/api/spots/wspr', async (req, res) => {
-  try {
-    if (wsprCache.data && Date.now() < wsprCache.expires) {
-      return res.json(wsprCache.data);
-    }
-
-    const sql = `SELECT time, tx_sign, rx_sign, frequency, snr, power,
-                        distance, drift, tx_lat, tx_lon, rx_lat, rx_lon, tx_loc, rx_loc
-                 FROM wspr.rx
-                 WHERE time > now() - INTERVAL 30 MINUTE
-                 ORDER BY time DESC
-                 LIMIT 1000
-                 FORMAT JSON`;
-    const url = `https://db1.wspr.live/?query=${encodeURIComponent(sql)}`;
-    const json = await fetchJSON(url);
-    const rows = json.data || [];
-
-    const spots = rows.map(r => {
-      const freqMHz = (parseInt(r.frequency, 10) / 1e6).toFixed(4);
-      const snrVal = parseInt(r.snr, 10);
-      const snrStr = isNaN(snrVal) ? '' : `${snrVal > 0 ? '+' : ''}${snrVal}`;
-      return {
-        callsign: r.tx_sign || '',
-        rxSign: r.rx_sign || '',
-        frequency: freqMHz,
-        band: freqToBandStr(parseFloat(freqMHz)),
-        mode: 'WSPR',
-        snr: snrStr,
-        power: `${r.power} dBm`,
-        distance: r.distance ? `${r.distance}` : '',
-        drift: r.drift,
-        latitude: parseFloat(r.tx_lat) || null,
-        longitude: parseFloat(r.tx_lon) || null,
-        rxLat: parseFloat(r.rx_lat) || null,
-        rxLon: parseFloat(r.rx_lon) || null,
-        name: `RX: ${r.rx_sign || '?'}`,
-        comments: `SNR: ${snrStr || '?'} dB · Power: ${r.power} dBm · Dist: ${r.distance || '?'} km`,
-        spotTime: r.time || '',
-        reference: '',
-      };
-    });
-
-    // Filter out spots with no valid transmitter coordinates
-    const validSpots = spots.filter(s => s.latitude !== null && s.longitude !== null);
-
-    wsprCache = { data: validSpots, expires: Date.now() + WSPR_CACHE_TTL };
-    res.json(validSpots);
-  } catch (err) {
-    console.error('Error fetching WSPR spots:', err.message);
-    res.status(502).json({ error: 'Failed to fetch WSPR spots' });
-  }
-});
-
-// Proxy SOTA spots API
-app.get('/api/spots/sota', async (req, res) => {
-  try {
-    const data = await fetchJSON('https://api2.sota.org.uk/api/spots/50/all');
-    const spots = (Array.isArray(data) ? data : []).map(s => ({
-      callsign:    s.activatorCallsign || '',
-      frequency:   s.frequency ? String(s.frequency) : '',
-      mode:        s.mode || '',
-      reference:   s.associationCode && s.summitCode
-                     ? s.associationCode + '/' + s.summitCode : (s.summitCode || ''),
-      name:        s.summitDetails || '',
-      spotTime:    s.timeStamp || '',
-      comments:    s.comments || '',
-      points:      s.points || 0,
-      latitude:    null,
-      longitude:   null,
-      _assoc:      s.associationCode || '',
-      _summit:     s.summitCode || '',
-    }));
-
-    // Look up coordinates for each unique summit
-    const unique = new Map();
-    for (const s of spots) {
-      if (s._assoc && s._summit) {
-        const key = `${s._assoc}/${s._summit}`;
-        if (!unique.has(key)) unique.set(key, { assoc: s._assoc, summit: s._summit });
-      }
-    }
-    await Promise.allSettled(
-      [...unique.values()].map(u => fetchSummitCoords(u.assoc, u.summit))
-    );
-
-    // Merge coordinates and strip internal fields
-    const enriched = spots.map(s => {
-      if (s._assoc && s._summit) {
-        const coords = sotaSummitCache[`${s._assoc}/${s._summit}`];
-        if (coords) {
-          s.latitude = coords.lat;
-          s.longitude = coords.lon;
-        }
-      }
-      delete s._assoc;
-      delete s._summit;
-      return s;
-    });
-
-    res.json(enriched);
-  } catch (err) {
-    console.error('Error fetching SOTA spots:', err.message);
-    res.status(502).json({ error: 'Failed to fetch SOTA spots' });
-  }
-});
-
-// --- WWFF (World Wide Flora & Fauna) Spots API ---
-
-app.get('/api/spots/wwff', async (req, res) => {
-  try {
-    const data = await fetchJSON('https://spots.wwff.co/static/spots.json');
-    const spots = (Array.isArray(data) ? data : []).map(s => ({
-      callsign:   s.activator || '',
-      frequency:  s.frequency_khz ? String(s.frequency_khz / 1000) : '', // kHz → MHz
-      mode:       s.mode || '',
-      reference:  s.reference || '',
-      name:       s.reference_name || '',
-      spotTime:   s.spot_time ? new Date(s.spot_time * 1000).toISOString() : '', // Unix → ISO
-      comments:   s.remarks || '',
-      latitude:   s.latitude ?? null,
-      longitude:  s.longitude ?? null,
-    }));
-    res.json(spots);
-  } catch (err) {
-    console.error('Error fetching WWFF spots:', err.message);
-    res.status(502).json({ error: 'Failed to fetch WWFF spots' });
-  }
-});
-
-// Fetch and parse solar XML data
-app.get('/api/solar', async (req, res) => {
-  try {
-    const xml = await fetchText('https://www.hamqsl.com/solarxml.php');
-    const solar = parseSolarXML(xml);
-    res.json(solar);
-  } catch (err) {
-    console.error('Error fetching solar data:', err.message);
-    res.status(502).json({ error: 'Failed to fetch solar data' });
-  }
-});
-
-// --- Space Weather History (NOAA SWPC) ---
-
-// Per-type in-memory cache, 15-minute TTL
-const spacewxCache = {}; // { type: { data, expires } }
-const SPACEWX_TTL = 15 * 60 * 1000; // 15 minutes
-
-const SPACEWX_URLS = {
-  kp:   'https://services.swpc.noaa.gov/products/noaa-planetary-k-index.json',
-  xray: 'https://services.swpc.noaa.gov/json/goes/primary/xrays-7-day.json',
-  sfi:  'https://services.swpc.noaa.gov/json/f107_cm_flux.json',
-  wind: 'https://services.swpc.noaa.gov/products/solar-wind/plasma-7-day.json',
-  mag:  'https://services.swpc.noaa.gov/products/solar-wind/mag-7-day.json',
-};
-
-// Downsample high-resolution data by bucket-averaging into ~targetPts points
-function downsampleSpacewx(arr, targetPts) {
-  if (arr.length <= targetPts) return arr;
-  const bucketSize = Math.ceil(arr.length / targetPts);
-  const result = [];
-  for (let i = 0; i < arr.length; i += bucketSize) {
-    const bucket = arr.slice(i, i + bucketSize);
-    const avgTime = bucket[Math.floor(bucket.length / 2)].time_tag;
-    // Average each numeric field in the bucket
-    const keys = Object.keys(bucket[0]).filter(k => k !== 'time_tag');
-    const avg = { time_tag: avgTime };
-    for (const k of keys) {
-      const vals = bucket.map(b => b[k]).filter(v => v !== null && !isNaN(v));
-      avg[k] = vals.length > 0 ? vals.reduce((a, b) => a + b, 0) / vals.length : null;
-    }
-    result.push(avg);
-  }
-  return result;
-}
-
-// Normalize NOAA data into uniform { time_tag, value, [value2] } format
-function normalizeSpacewx(type, raw) {
-  switch (type) {
-    case 'kp': {
-      // First row is header: ["time_tag","Kp","Kp_fraction",...]
-      const rows = raw.slice(1);
-      return rows.map(r => ({
-        time_tag: r[0],
-        value: parseFloat(r[1]),
-      })).filter(r => !isNaN(r.value));
-    }
-    case 'xray': {
-      // Array of objects: { time_tag, flux }
-      const parsed = raw.map(r => ({
-        time_tag: r.time_tag,
-        value: parseFloat(r.flux),
-      })).filter(r => !isNaN(r.value) && r.value > 0);
-      return downsampleSpacewx(parsed, 500);
-    }
-    case 'sfi': {
-      // Array of objects: { time_tag, flux }
-      return raw.map(r => ({
-        time_tag: r.time_tag,
-        value: parseFloat(r.flux),
-      })).filter(r => !isNaN(r.value));
-    }
-    case 'wind': {
-      // First row is header: ["time_tag","density","speed","temperature"]
-      const rows = raw.slice(1);
-      const parsed = rows.map(r => ({
-        time_tag: r[0],
-        value: parseFloat(r[2]), // speed column
-      })).filter(r => !isNaN(r.value));
-      return downsampleSpacewx(parsed, 500);
-    }
-    case 'mag': {
-      // First row is header: ["time_tag","bx_gsm","by_gsm","bz_gsm","lon_gsm","lat_gsm","bt"]
-      const rows = raw.slice(1);
-      const parsed = rows.map(r => ({
-        time_tag: r[0],
-        value: parseFloat(r[3]),  // bz_gsm
-        value2: parseFloat(r[6]), // bt (total field magnitude)
-      })).filter(r => !isNaN(r.value));
-      return downsampleSpacewx(parsed, 500);
-    }
-    default:
-      return [];
-  }
-}
-
-app.get('/api/spacewx/history', async (req, res) => {
-  const type = req.query.type;
-  if (!type || !SPACEWX_URLS[type]) {
-    return res.status(400).json({ error: 'Invalid type. Use: kp, xray, sfi, wind, mag' });
-  }
-
-  try {
-    // Return cached data if fresh
-    const cached = spacewxCache[type];
-    if (cached && Date.now() < cached.expires) {
-      return res.json(cached.data);
-    }
-
-    const raw = await fetchJSON(SPACEWX_URLS[type]);
-    const data = normalizeSpacewx(type, raw);
-
-    spacewxCache[type] = { data, expires: Date.now() + SPACEWX_TTL };
-    res.json(data);
-  } catch (err) {
-    console.error(`Error fetching spacewx ${type}:`, err.message);
-    res.status(502).json({ error: `Failed to fetch space weather ${type} data` });
-  }
-});
-
-// --- N2YO Satellite Tracking API ---
-
-// Satellite list cache (category 18 = amateur radio)
-let satListCache = { data: null, expires: 0 };
-const SAT_LIST_TTL = 24 * 60 * 60 * 1000; // 24 hours
-
-// Satellite position cache
-let satPosCache = { data: null, expires: 0, key: '' };
-const SAT_POS_TTL = 10 * 1000; // 10 seconds
-
-// Pass prediction cache (per satellite)
-const satPassCache = {}; // { 'satId:lat:lon': { data, expires } }
-const SAT_PASS_TTL = 5 * 60 * 1000; // 5 minutes
-const SAT_PASS_MAX = 500; // max cached entries
-
-// TLE cache (per satellite)
-const satTleCache = {}; // { satId: { data, expires } }
-const SAT_TLE_TTL = 6 * 60 * 60 * 1000; // 6 hours
-const SAT_TLE_MAX = 500; // max cached entries
-
-// Fetch amateur radio satellite list (N2YO category 18)
-app.get('/api/satellites/list', async (req, res) => {
-  const apiKey = req.headers['x-api-key'] || req.query.apikey || process.env.N2YO_API_KEY;
-  if (!apiKey) {
-    return res.status(503).json({ error: 'No N2YO API key configured' });
-  }
-
-  try {
-    // Return cached data if fresh
-    if (satListCache.data && Date.now() < satListCache.expires) {
-      return res.json(satListCache.data);
-    }
-
-    const url = `https://api.n2yo.com/rest/v1/satellite/above/0/0/0/90/18/&apiKey=${apiKey}`;
-    const data = await fetchJSON(url);
-
-    // Extract satellite list from response
-    const satellites = (data.above || []).map(s => ({
-      satId: s.satid,
-      name: s.satname,
-      intDesignator: s.intDesignator,
-    }));
-
-    satListCache = { data: satellites, expires: Date.now() + SAT_LIST_TTL };
-    res.json(satellites);
-  } catch (err) {
-    console.error('Error fetching satellite list:', err.message);
-    res.status(502).json({ error: 'Failed to fetch satellite list' });
-  }
-});
-
-// Fetch positions for multiple satellites (batched)
-app.get('/api/satellites/positions', async (req, res) => {
-  const apiKey = req.headers['x-api-key'] || req.query.apikey || process.env.N2YO_API_KEY;
-  if (!apiKey) {
-    return res.status(503).json({ error: 'No N2YO API key configured' });
-  }
-
-  const { ids, lat, lon, seconds } = req.query;
-  if (!ids) {
-    return res.status(400).json({ error: 'Provide satellite IDs' });
-  }
-
-  const obsLat = parseFloat(lat) || 0;
-  const obsLon = parseFloat(lon) || 0;
-  if (obsLat < -90 || obsLat > 90 || obsLon < -180 || obsLon > 180) {
-    return res.status(400).json({ error: 'lat must be -90..90, lon must be -180..180' });
-  }
-  const secs = parseInt(seconds, 10) || 1;
-  const satIds = ids.split(',').map(id => id.trim()).filter(Boolean).slice(0, 10); // max 10 satellites
-
-  if (satIds.length === 0) {
-    return res.status(400).json({ error: 'Provide at least one satellite ID' });
-  }
-  if (!satIds.every(id => /^\d+$/.test(id))) {
-    return res.status(400).json({ error: 'Satellite IDs must be numeric' });
-  }
-
-  // Cache key includes IDs + rounded observer coords (~1.1 km granularity)
-  const cacheKey = `${satIds.sort().join(',')}:${obsLat.toFixed(2)}:${obsLon.toFixed(2)}`;
-  if (satPosCache.key === cacheKey && satPosCache.data && Date.now() < satPosCache.expires) {
-    return res.json(satPosCache.data);
-  }
-
-  try {
-    const positions = {};
-
-    // Fetch positions for each satellite in parallel
-    await Promise.all(satIds.map(async (satId) => {
-      try {
-        const url = `https://api.n2yo.com/rest/v1/satellite/positions/${encodeURIComponent(satId)}/${obsLat}/${obsLon}/0/${secs}/&apiKey=${apiKey}`;
-        const data = await fetchJSON(url);
-        if (data.positions && data.positions.length > 0) {
-          const pos = data.positions[0];
-          positions[satId] = {
-            satId: data.info?.satid || parseInt(satId, 10),
-            name: data.info?.satname || '',
-            lat: pos.satlatitude,
-            lon: pos.satlongitude,
-            alt: pos.sataltitude,
-            azimuth: pos.azimuth,
-            elevation: pos.elevation,
-            ra: pos.ra,
-            dec: pos.dec,
-            timestamp: pos.timestamp,
-          };
-        }
-      } catch (err) {
-        console.error(`Error fetching position for sat ${satId}:`, err.message);
-      }
-    }));
-
-    satPosCache = { data: positions, expires: Date.now() + SAT_POS_TTL, key: cacheKey };
-    res.json(positions);
-  } catch (err) {
-    console.error('Error fetching satellite positions:', err.message);
-    res.status(502).json({ error: 'Failed to fetch satellite positions' });
-  }
-});
-
-// Fetch pass predictions for a single satellite
-app.get('/api/satellites/passes', async (req, res) => {
-  const apiKey = req.headers['x-api-key'] || req.query.apikey || process.env.N2YO_API_KEY;
-  if (!apiKey) {
-    return res.status(503).json({ error: 'No N2YO API key configured' });
-  }
-
-  const { id, lat, lon, days, minEl } = req.query;
-  if (!id || !/^\d+$/.test(id)) {
-    return res.status(400).json({ error: 'Provide a valid numeric satellite ID' });
-  }
-
-  const obsLat = parseFloat(lat);
-  const obsLon = parseFloat(lon);
-  if (isNaN(obsLat) || isNaN(obsLon) || obsLat < -90 || obsLat > 90 || obsLon < -180 || obsLon > 180) {
-    return res.status(400).json({ error: 'Provide valid lat (-90..90) and lon (-180..180)' });
-  }
-
-  const numDays = Math.min(parseInt(days, 10) || 2, 10); // max 10 days
-  const minElevation = Math.min(parseInt(minEl, 10) || 10, 90);
-
-  // Check cache
-  const cacheKey = `${id}:${obsLat.toFixed(2)}:${obsLon.toFixed(2)}`;
-  const cached = satPassCache[cacheKey];
-  if (cached && Date.now() < cached.expires) {
-    return res.json(cached.data);
-  }
-
-  try {
-    const url = `https://api.n2yo.com/rest/v1/satellite/radiopasses/${encodeURIComponent(id)}/${obsLat}/${obsLon}/0/${numDays}/${minElevation}/&apiKey=${apiKey}`;
-    const data = await fetchJSON(url);
-
-    const passes = (data.passes || []).map(p => ({
-      startUTC: p.startUTC,
-      startAz: p.startAz,
-      startAzCompass: p.startAzCompass,
-      maxUTC: p.maxUTC,
-      maxAz: p.maxAz,
-      maxAzCompass: p.maxAzCompass,
-      maxEl: p.maxEl,
-      endUTC: p.endUTC,
-      endAz: p.endAz,
-      endAzCompass: p.endAzCompass,
-    }));
-
-    const result = {
-      satId: data.info?.satid,
-      name: data.info?.satname,
-      passesCount: data.info?.passescount || passes.length,
-      passes,
-    };
-
-    if (Object.keys(satPassCache).length >= SAT_PASS_MAX) {
-      const now = Date.now();
-      for (const k of Object.keys(satPassCache)) {
-        if (satPassCache[k].expires < now) delete satPassCache[k];
-      }
-    }
-    satPassCache[cacheKey] = { data: result, expires: Date.now() + SAT_PASS_TTL };
-    res.json(result);
-  } catch (err) {
-    console.error('Error fetching satellite passes:', err.message);
-    res.status(502).json({ error: 'Failed to fetch satellite passes' });
-  }
-});
-
-// Fetch TLE data for a satellite (used for Doppler calculations)
-app.get('/api/satellites/tle', async (req, res) => {
-  const apiKey = req.headers['x-api-key'] || req.query.apikey || process.env.N2YO_API_KEY;
-  if (!apiKey) {
-    return res.status(503).json({ error: 'No N2YO API key configured' });
-  }
-
-  const { id } = req.query;
-  if (!id || !/^\d+$/.test(id)) {
-    return res.status(400).json({ error: 'Provide a valid numeric satellite ID' });
-  }
-
-  // Check cache
-  const cached = satTleCache[id];
-  if (cached && Date.now() < cached.expires) {
-    return res.json(cached.data);
-  }
-
-  try {
-    const url = `https://api.n2yo.com/rest/v1/satellite/tle/${encodeURIComponent(id)}/&apiKey=${apiKey}`;
-    const data = await fetchJSON(url);
-
-    const result = {
-      satId: data.info?.satid,
-      name: data.info?.satname,
-      tle: data.tle,
-    };
-
-    if (Object.keys(satTleCache).length >= SAT_TLE_MAX) {
-      const now = Date.now();
-      for (const k of Object.keys(satTleCache)) {
-        if (satTleCache[k].expires < now) delete satTleCache[k];
-      }
-    }
-    satTleCache[id] = { data: result, expires: Date.now() + SAT_TLE_TTL };
-    res.json(result);
-  } catch (err) {
-    console.error('Error fetching satellite TLE:', err.message);
-    res.status(502).json({ error: 'Failed to fetch satellite TLE' });
-  }
-});
-
-// --- Keyless ISS Tracking (SGP4 via satellite.js) ---
-
-// ISS TLE cache — fetched from CelesTrak, no API key needed
-let issTleCache = { data: null, expires: 0 };
-const ISS_TLE_TTL = 6 * 60 * 60 * 1000; // 6h — TLE doesn't change often
-
-// ISS computed position/orbit cache — recomputed every 10s
-let issComputedCache = { data: null, expires: 0, obsKey: '' };
-const ISS_COMPUTED_TTL = 10 * 1000; // 10s — matches satellite position refresh
-
-// Fetch ISS TLE from CelesTrak (free, no API key)
-async function fetchIssTle() {
-  if (issTleCache.data && Date.now() < issTleCache.expires) {
-    return issTleCache.data;
-  }
-
-  const url = 'https://celestrak.org/NORAD/elements/gp.php?CATNR=25544&FORMAT=tle';
-  const text = await fetchText(url);
-  const lines = text.trim().split('\n').map(l => l.trim());
-
-  if (lines.length < 3) {
-    throw new Error('Invalid TLE response from CelesTrak');
-  }
-
-  // 3-line TLE format: name, line1, line2
-  const tle = { name: lines[0], line1: lines[1], line2: lines[2] };
-  issTleCache = { data: tle, expires: Date.now() + ISS_TLE_TTL };
-  return tle;
-}
-
-// Compute ISS orbit ground track — 100 points over ~92.5 min (one full orbit)
-// Starts 10 min in the past to show where ISS came from
-function computeIssOrbitPath(satrec) {
-  const points = [];
-  const now = new Date();
-  const startOffset = -10 * 60 * 1000; // 10 min ago
-  const endOffset = 82.5 * 60 * 1000;  // 82.5 min ahead (total ~92.5 min = 1 orbit)
-  const step = (endOffset - startOffset) / 99; // 100 points
-
-  for (let i = 0; i < 100; i++) {
-    const t = new Date(now.getTime() + startOffset + step * i);
-    const posVel = satellite.propagate(satrec, t);
-    if (!posVel.position) continue;
-
-    const gmst = satellite.gstime(t);
-    const geo = satellite.eciToGeodetic(posVel.position, gmst);
-    const lat = satellite.degreesLat(geo.latitude);
-    const lon = satellite.degreesLong(geo.longitude);
-
-    if (!isNaN(lat) && !isNaN(lon)) {
-      points.push({ lat: parseFloat(lat.toFixed(4)), lon: parseFloat(lon.toFixed(4)) });
-    }
-  }
-
-  return points;
-}
-
-// ISS position endpoint — free, no API key required
-app.get('/api/iss/position', async (req, res) => {
-  const obsLat = parseFloat(req.query.lat) || 0;
-  const obsLon = parseFloat(req.query.lon) || 0;
-  if (obsLat < -90 || obsLat > 90 || obsLon < -180 || obsLon > 180) {
-    return res.status(400).json({ error: 'lat must be -90..90, lon must be -180..180' });
-  }
-  const obsKey = `${obsLat.toFixed(2)}:${obsLon.toFixed(2)}`;
-
-  // Return cached data if fresh and same observer
-  if (issComputedCache.data && Date.now() < issComputedCache.expires && issComputedCache.obsKey === obsKey) {
-    return res.json(issComputedCache.data);
-  }
-
-  try {
-    const tle = await fetchIssTle();
-    const satrec = satellite.twoline2satrec(tle.line1, tle.line2);
-
-    // Current position
-    const now = new Date();
-    const posVel = satellite.propagate(satrec, now);
-    if (!posVel.position) {
-      return res.status(500).json({ error: 'SGP4 propagation failed' });
-    }
-
-    const gmst = satellite.gstime(now);
-    const geo = satellite.eciToGeodetic(posVel.position, gmst);
-    const lat = satellite.degreesLat(geo.latitude);
-    const lon = satellite.degreesLong(geo.longitude);
-    const alt = geo.height; // km
-
-    // Velocity magnitude (km/s) from ECI velocity vector
-    const vx = posVel.velocity.x, vy = posVel.velocity.y, vz = posVel.velocity.z;
-    const velocity = Math.sqrt(vx * vx + vy * vy + vz * vz);
-
-    // Observer look angles (azimuth/elevation)
-    let azimuth = 0, elevation = -90;
-    if (obsLat !== 0 || obsLon !== 0) {
-      const obsGeo = {
-        longitude: satellite.degreesToRadians(obsLon),
-        latitude: satellite.degreesToRadians(obsLat),
-        height: 0, // km above sea level
-      };
-      const obsEcf = satellite.geodeticToEcf(obsGeo);
-      const lookAngles = satellite.ecfToLookAngles(obsGeo, satellite.eciToEcf(posVel.position, gmst));
-      azimuth = satellite.radiansToDegrees(lookAngles.azimuth);
-      elevation = satellite.radiansToDegrees(lookAngles.elevation);
-    }
-
-    // Orbit path (100 points over ~1 orbit)
-    const orbitPath = computeIssOrbitPath(satrec);
-
-    // TLE epoch — satellite.js stores epochyr (2-digit year) and epochdays (fractional day of year)
-    const epochYear = satrec.epochyr < 57 ? 2000 + satrec.epochyr : 1900 + satrec.epochyr;
-    const epochDate = new Date(Date.UTC(epochYear, 0, 1));
-    epochDate.setUTCDate(epochDate.getUTCDate() + Math.floor(satrec.epochdays) - 1);
-    const tleEpoch = Math.floor(epochDate.getTime() / 1000); // Unix timestamp (seconds)
-
-    const result = {
-      satId: 25544,
-      name: 'ISS (ZARYA)',
-      lat: parseFloat(lat.toFixed(4)),
-      lon: parseFloat(lon.toFixed(4)),
-      alt: parseFloat(alt.toFixed(1)),
-      azimuth: parseFloat(azimuth.toFixed(1)),
-      elevation: parseFloat(elevation.toFixed(1)),
-      velocity: parseFloat(velocity.toFixed(2)),
-      timestamp: Math.floor(now.getTime() / 1000),
-      tleEpoch,
-      orbitPath,
-    };
-
-    issComputedCache = { data: result, expires: Date.now() + ISS_COMPUTED_TTL, obsKey };
-    res.json(result);
-  } catch (err) {
-    console.error('Error computing ISS position:', err.message);
-    res.status(502).json({ error: 'Failed to compute ISS position' });
-  }
-});
-
-// Lunar / EME conditions
-app.get('/api/lunar', (req, res) => {
-  try {
-    // Optional observer coordinates for az/el and rise/set
-    const lat = parseFloat(req.query.lat);
-    const lon = parseFloat(req.query.lon);
-    const hasObserver = !isNaN(lat) && !isNaN(lon)
-      && lat >= -90 && lat <= 90 && lon >= -180 && lon <= 180;
-    const lunar = hasObserver ? computeLunar(lat, lon) : computeLunar();
-    res.json(lunar);
-  } catch (err) {
-    console.error('Error computing lunar data:', err.message);
-    res.status(500).json({ error: 'Failed to compute lunar data' });
-  }
-});
-
-// (Weather routes moved to server/routes/weather.js)
-
-// (Weather conditions/alerts/OWM/NWS fetch moved to server/routes/weather.js)
-// (Callsign route moved to server/routes/callsign.js)
-
-// --- Mode-Specific Endpoints ---
-// (Empty on main - populated on deployment branches)
-// Lanmode adds: /api/update/*, /api/restart
-// Hostedmode adds: /api/settings-sync (future)
-
-// (Config endpoint moved to server/routes/config.js)
-
-// (Feedback endpoint moved to server/routes/config.js)
-
-// Proxy NASA SDO solar images
-const SDO_TYPES = new Set(['0193', '0171', '0304', 'HMIIC']);
-
-// --- SDO browse frame list (animated time-lapse) ---
-// Cache per-day directory listings to avoid re-scraping
-const sdoDirCache = {}; // { 'YYYY/MM/DD:type': { frames: [], expires: timestamp } }
-
-// Scrape one day's directory listing for frame filenames
-async function scrapeSDODay(dateKey, type) {
-  const cacheKey = `${dateKey}:${type}`;
-  const cached = sdoDirCache[cacheKey];
-  if (cached && Date.now() < cached.expires) return cached.frames;
-
-  const html = await fetchText(`https://sdo.gsfc.nasa.gov/assets/img/browse/${dateKey}/`);
-  const pattern = new RegExp(`\\d{8}_\\d{6}_512_${type}\\.jpg`, 'g');
-  const frames = [];
-  let match;
-  while ((match = pattern.exec(html)) !== null) frames.push(match[0]);
-
-  // Yesterday's listing is immutable — cache for 6 hours; today's for 10 minutes
-  const now = new Date();
-  const todayKey = `${now.getUTCFullYear()}/${String(now.getUTCMonth() + 1).padStart(2, '0')}/${String(now.getUTCDate()).padStart(2, '0')}`;
-  const ttl = (dateKey === todayKey) ? 10 * 60 * 1000 : 6 * 3600 * 1000;
-  sdoDirCache[cacheKey] = { frames, expires: Date.now() + ttl };
-  return frames;
-}
-
-app.get('/api/solar/frames', async (req, res) => {
-  try {
-    const type = SDO_TYPES.has(req.query.type) ? req.query.type : '0193';
-    const now = new Date();
-    const todayKey = `${now.getUTCFullYear()}/${String(now.getUTCMonth() + 1).padStart(2, '0')}/${String(now.getUTCDate()).padStart(2, '0')}`;
-
-    // Yesterday's date
-    const yesterday = new Date(now);
-    yesterday.setUTCDate(yesterday.getUTCDate() - 1);
-    const yesterdayKey = `${yesterday.getUTCFullYear()}/${String(yesterday.getUTCMonth() + 1).padStart(2, '0')}/${String(yesterday.getUTCDate()).padStart(2, '0')}`;
-
-    // Fetch both days in parallel
-    const [yesterdayFrames, todayFrames] = await Promise.all([
-      scrapeSDODay(yesterdayKey, type).catch(() => []),
-      scrapeSDODay(todayKey, type).catch(() => []),
-    ]);
-
-    // Combine and return last 48 frames (~8 hours of coverage)
-    const allFrames = [...yesterdayFrames, ...todayFrames];
-    const frames = allFrames.slice(-48);
-    res.json(frames);
-  } catch (err) {
-    console.error('Error fetching SDO frame list:', err.message);
-    res.status(502).json({ error: 'Failed to fetch SDO frame list' });
-  }
-});
-
-// Proxy individual SDO browse frames
-const SDO_FRAME_RE = /^\d{8}_\d{6}_512_\w+\.jpg$/;
-
-app.get('/api/solar/frame/:filename', async (req, res) => {
-  try {
-    const filename = req.params.filename;
-    if (!SDO_FRAME_RE.test(filename)) {
-      return res.status(400).json({ error: 'Invalid filename' });
-    }
-
-    // Extract date from filename (YYYYMMDD_HHMMSS_512_type.jpg)
-    const yyyy = filename.substring(0, 4);
-    const mm = filename.substring(4, 6);
-    const dd = filename.substring(6, 8);
-    const url = `https://sdo.gsfc.nasa.gov/assets/img/browse/${yyyy}/${mm}/${dd}/${filename}`;
-    const parsed = new URL(url);
-    const resolvedIP = await resolveHost(parsed.hostname);
-    if (isPrivateIP(resolvedIP)) {
-      return res.status(403).json({ error: 'Blocked' });
-    }
-    const proxyReq = https.get({
-      hostname: resolvedIP,
-      path: parsed.pathname,
-      port: 443,
-      headers: { 'User-Agent': 'HamTab/1.0', 'Host': parsed.hostname },
-      servername: parsed.hostname,
-    }, (upstream) => {
-      if (upstream.statusCode < 200 || upstream.statusCode >= 300) {
-        upstream.resume();
-        return res.status(502).json({ error: 'Failed to fetch SDO frame' });
-      }
-      res.set('Content-Type', upstream.headers['content-type'] || 'image/jpeg');
-      res.set('Cache-Control', 'public, max-age=86400');
-      upstream.pipe(res);
-    });
-    proxyReq.on('error', (err) => {
-      console.error('SDO frame proxy error:', err.message);
-      if (!res.headersSent) res.status(502).json({ error: 'Failed to fetch SDO frame' });
-    });
-    proxyReq.setTimeout(15000, () => { proxyReq.destroy(); });
-  } catch (err) {
-    console.error('Error fetching SDO frame:', err.message);
-    res.status(502).json({ error: 'Failed to fetch SDO frame' });
-  }
-});
-
-app.get('/api/solar/image', async (req, res) => {
-  try {
-    const type = SDO_TYPES.has(req.query.type) ? req.query.type : '0193';
-    const url = `https://sdo.gsfc.nasa.gov/assets/img/latest/latest_512_${type}.jpg`;
-    const parsed = new URL(url);
-    const resolvedIP = await resolveHost(parsed.hostname);
-    if (isPrivateIP(resolvedIP)) {
-      return res.status(403).json({ error: 'Blocked' });
-    }
-    const proxyReq = https.get({
-      hostname: resolvedIP,
-      path: parsed.pathname,
-      port: 443,
-      headers: { 'User-Agent': 'HamTab/1.0', 'Host': parsed.hostname },
-      servername: parsed.hostname,
-    }, (upstream) => {
-      if (upstream.statusCode < 200 || upstream.statusCode >= 300) {
-        upstream.resume();
-        return res.status(502).json({ error: 'Failed to fetch SDO image' });
-      }
-      res.set('Content-Type', upstream.headers['content-type'] || 'image/jpeg');
-      res.set('Cache-Control', 'public, max-age=300');
-      upstream.pipe(res);
-    });
-    proxyReq.on('error', (err) => {
-      console.error('SDO image proxy error:', err.message);
-      if (!res.headersSent) res.status(502).json({ error: 'Failed to fetch SDO image' });
-    });
-    proxyReq.setTimeout(15000, () => { proxyReq.destroy(); });
-  } catch (err) {
-    console.error('Error fetching SDO image:', err.message);
-    res.status(502).json({ error: 'Failed to fetch SDO image' });
-  }
-});
-
-// Proxy NASA SVS moon phase image (pre-rendered LROC frames for current year)
-// Frame number = hour of year (1–8760), gives accurate texture + libration + lighting
-app.get('/api/lunar/image', async (req, res) => {
-  try {
-    const now = new Date();
-    const startOfYear = new Date(Date.UTC(now.getUTCFullYear(), 0, 1));
-    const hourOfYear = Math.floor((now - startOfYear) / 3600000) + 1;
-    const frame = String(Math.min(hourOfYear, 8760)).padStart(4, '0');
-    const url = `https://svs.gsfc.nasa.gov/vis/a000000/a005500/a005587/frames/730x730_1x1_30p/moon.${frame}.jpg`;
-    const parsed = new URL(url);
-    const resolvedIP = await resolveHost(parsed.hostname);
-    if (isPrivateIP(resolvedIP)) {
-      return res.status(403).json({ error: 'Blocked' });
-    }
-    const proxyReq = https.get({
-      hostname: resolvedIP,
-      path: parsed.pathname,
-      port: 443,
-      headers: { 'User-Agent': 'HamTab/1.0', 'Host': parsed.hostname },
-      servername: parsed.hostname,
-    }, (upstream) => {
-      if (upstream.statusCode < 200 || upstream.statusCode >= 300) {
-        upstream.resume();
-        return res.status(502).json({ error: 'Failed to fetch moon image' });
-      }
-      res.set('Content-Type', upstream.headers['content-type'] || 'image/jpeg');
-      // Cache 1 hour — frame changes hourly
-      res.set('Cache-Control', 'public, max-age=3600');
-      upstream.pipe(res);
-    });
-    proxyReq.on('error', (err) => {
-      console.error('Moon image proxy error:', err.message);
-      if (!res.headersSent) res.status(502).json({ error: 'Failed to fetch moon image' });
-    });
-    proxyReq.setTimeout(15000, () => { proxyReq.destroy(); });
-  } catch (err) {
-    console.error('Error fetching moon image:', err.message);
-    res.status(502).json({ error: 'Failed to fetch moon image' });
-  }
-});
-
-// --- DXpeditions (NG3K ADXO feed) ---
-
-const dxpeditionCache = { data: null, expires: 0 };
-const DXPEDITION_TTL = 2 * 60 * 60 * 1000; // 2 hours
-
-// DXCC entity name → approximate [lat, lon] for DXpedition map markers.
-// Keys are lowercase. Covers ~300 current DXCC entities + common NG3K name variants.
-const DXCC_COORDS = {
-  'afghanistan': [34.5, 69.2], 'agalega & st brandon': [-10.4, 56.6], 'aland is.': [60.2, 20.0],
-  'aland islands': [60.2, 20.0], 'alaska': [64.2, -152.5], 'albania': [41.3, 20.0],
-  'algeria': [28.0, 3.0], 'american samoa': [-14.3, -170.7], 'amsterdam & st paul': [-37.8, 77.6],
-  'amsterdam & st. paul': [-37.8, 77.6], 'andaman & nicobar': [12.0, 92.8],
-  'andorra': [42.5, 1.5], 'angola': [-12.5, 18.5], 'anguilla': [18.2, -63.1],
-  'annobon': [-1.4, 5.6], 'annobon island': [-1.4, 5.6], 'antarctica': [-82.0, 0.0],
-  'antigua & barbuda': [17.1, -61.8], 'argentina': [-34.6, -58.4], 'armenia': [40.2, 44.5],
-  'aruba': [12.5, -70.0], 'ascension island': [-7.9, -14.4], 'asiatic russia': [60.0, 100.0],
-  'asiatic turkey': [39.9, 32.9], 'australia': [-25.3, 133.8], 'austria': [47.5, 13.3],
-  'aves island': [15.7, -63.6], 'azerbaijan': [40.4, 49.9], 'azores': [38.7, -27.2],
-  'bahamas': [25.0, -77.4], 'bahrain': [26.0, 50.6], 'baker & howland is.': [0.2, -176.5],
-  'balearic islands': [39.6, 2.9], 'bangladesh': [23.7, 90.4], 'barbados': [13.2, -59.5],
-  'barbuda': [17.6, -61.8], 'belarus': [53.9, 27.6], 'belgium': [50.8, 4.4],
-  'belize': [17.2, -88.7], 'benin': [9.3, 2.3], 'bermuda': [32.3, -64.8],
-  'bhutan': [27.5, 90.4], 'bolivia': [-16.5, -68.2], 'bonaire': [12.1, -68.3],
-  'bonaire, curacao': [12.1, -68.3], 'bosnia-herzegovina': [43.9, 17.7],
-  'botswana': [-22.3, 24.7], 'bouvet': [-54.4, 3.4], 'bouvet island': [-54.4, 3.4],
-  'brazil': [-15.8, -47.9], 'british virgin islands': [18.4, -64.6],
-  'brunei': [4.9, 114.9], 'bulgaria': [42.7, 25.5], 'burkina faso': [12.4, -1.5],
-  'burundi': [-3.4, 29.4], 'cambodia': [11.6, 104.9], 'cameroon': [3.9, 11.5],
-  'canada': [56.1, -106.3], 'canary islands': [28.1, -15.4], 'cape verde': [16.0, -24.0],
-  'cayman islands': [19.3, -81.3], 'central african republic': [4.4, 18.6],
-  'ceuta & melilla': [35.9, -5.3], 'chad': [12.1, 15.0], 'chagos islands': [-7.3, 72.4],
-  'chatham islands': [-44.0, -176.5], 'chile': [-33.4, -70.7], 'china': [35.9, 104.2],
-  'christmas island': [-10.5, 105.7], 'clipperton island': [10.3, -109.2],
-  'cocos (keeling)': [-12.2, 96.8], 'cocos island': [-12.2, 96.8],
-  'cocos-keeling': [-12.2, 96.8], 'colombia': [4.6, -74.1],
-  'comoros': [-12.2, 44.3], 'congo': [-4.3, 15.3], 'congo, dem. rep.': [-4.3, 15.3],
-  'congo, republic': [-0.2, 15.8], 'cook islands': [-21.2, -159.8],
-  'corsica': [42.0, 9.0], 'costa rica': [9.9, -84.1], 'cote d\'ivoire': [5.3, -4.0],
-  'crete': [35.2, 24.9], 'croatia': [45.1, 15.2], 'crozet island': [-46.4, 51.8],
-  'cuba': [21.5, -77.8], 'curacao': [12.2, -69.0], 'cyprus': [35.1, 33.4],
-  'czech republic': [49.8, 15.5], 'denmark': [56.3, 9.5], 'desecheo island': [18.4, -67.5],
-  'djibouti': [11.6, 43.1], 'dodecanese': [36.4, 28.0], 'dominica': [15.4, -61.4],
-  'dominican republic': [18.7, -70.2], 'ducie island': [-24.7, -124.8],
-  'east kiribati': [1.9, -157.5], 'east malaysia': [4.0, 114.0],
-  'east timor': [-8.6, 125.7], 'easter island': [-27.1, -109.4],
-  'ecuador': [-0.2, -78.5], 'egypt': [26.8, 30.8], 'el salvador': [13.8, -88.9],
-  'england': [51.5, -0.1], 'equatorial guinea': [1.9, 10.0], 'eritrea': [15.3, 38.9],
-  'estonia': [58.6, 25.0], 'eswatini': [-26.5, 31.5], 'ethiopia': [9.0, 38.7],
-  'european russia': [55.8, 37.6], 'european turkey': [41.0, 28.0],
-  'falkland islands': [-51.8, -59.0], 'faroe islands': [62.0, -6.8],
-  'fiji': [-18.1, 178.0], 'finland': [60.2, 24.9], 'fr. polynesia': [-17.7, -149.4],
-  'france': [46.2, 2.2], 'franz josef land': [80.0, 50.0],
-  'french guiana': [4.0, -53.0], 'french polynesia': [-17.7, -149.4],
-  'gabon': [-0.8, 11.6], 'galapagos islands': [-0.8, -91.1], 'gambia': [13.4, -16.6],
-  'georgia': [42.3, 43.4], 'germany': [51.2, 10.5], 'ghana': [5.6, -0.2],
-  'gibraltar': [36.1, -5.4], 'glorioso islands': [-11.6, 47.3],
-  'greece': [39.1, 21.8], 'greenland': [71.7, -42.6], 'grenada': [12.1, -61.7],
-  'guadeloupe': [16.3, -61.5], 'guam': [13.4, 144.8], 'guatemala': [14.6, -90.5],
-  'guernsey': [49.5, -2.5], 'guinea': [9.9, -13.7], 'guinea-bissau': [12.0, -15.0],
-  'guyana': [4.9, -58.9], 'haiti': [18.5, -72.3], 'hawaii': [19.9, -155.6],
-  'heard island': [-53.1, 73.5], 'honduras': [14.1, -87.2], 'hong kong': [22.3, 114.2],
-  'hungary': [47.5, 19.1], 'iceland': [65.0, -18.0], 'india': [20.6, 78.9],
-  'indonesia': [-6.2, 106.8], 'iran': [35.7, 51.4], 'iraq': [33.3, 44.4],
-  'ireland': [53.4, -8.2], 'isle of man': [54.2, -4.5], 'israel': [31.0, 34.8],
-  'italy': [41.9, 12.5], 'jamaica': [18.1, -77.3], 'jan mayen': [71.0, -8.5],
-  'japan': [36.2, 138.3], 'jarvis island': [-0.4, -160.0], 'jersey': [49.2, -2.1],
-  'johnston island': [16.7, -169.5], 'jordan': [30.6, 36.2],
-  'juan de nova & europa': [-17.1, 42.7], 'juan fernandez is.': [-33.6, -78.8],
-  'kaliningrad': [54.7, 20.5], 'kampuchea': [11.6, 104.9], 'kazakhstan': [48.0, 68.0],
-  'kenya': [-1.3, 36.8], 'kerguelen islands': [-49.4, 69.3],
-  'kermadec islands': [-29.3, -177.9], 'kingman reef': [6.4, -162.4],
-  'kiribati': [1.9, -157.5], 'korea': [37.6, 127.0], 'kosovo': [42.6, 20.9],
-  'kuwait': [29.4, 47.9], 'kure island': [28.4, -178.3], 'kyrgyzstan': [41.2, 74.8],
-  'lakshadweep islands': [10.1, 72.3], 'laos': [18.0, 102.6], 'latvia': [56.9, 24.1],
-  'lebanon': [33.9, 35.5], 'lesotho': [-29.6, 28.2], 'liberia': [6.4, -9.4],
-  'libya': [26.3, 17.2], 'liechtenstein': [47.1, 9.6], 'lithuania': [55.2, 23.9],
-  'lord howe island': [-31.6, 159.1], 'luxembourg': [49.8, 6.1],
-  'macao': [22.2, 113.5], 'macquarie island': [-54.6, 158.9],
-  'madagascar': [-18.8, 47.5], 'madeira islands': [32.6, -16.9], 'malawi': [-13.3, 34.3],
-  'malaysia': [3.1, 101.7], 'maldives': [3.2, 73.2], 'mali': [17.6, -4.0],
-  'malpelo island': [4.0, -81.6], 'malta': [35.9, 14.5], 'mariana islands': [15.2, 145.8],
-  'market reef': [60.3, 19.1], 'marquesas islands': [-9.8, -139.0],
-  'marshall islands': [7.1, 171.4], 'martinique': [14.6, -61.0], 'mauritania': [18.1, -15.9],
-  'mauritius': [-20.2, 57.5], 'mayotte': [-12.8, 45.2], 'mellish reef': [-17.4, 155.9],
-  'mexico': [23.6, -102.6], 'micronesia': [6.9, 158.2], 'midway island': [28.2, -177.4],
-  'minami torishima': [24.3, 154.0], 'moldova': [47.4, 28.4], 'monaco': [43.7, 7.4],
-  'mongolia': [47.9, 106.9], 'montenegro': [42.5, 19.3], 'montserrat': [16.7, -62.2],
-  'morocco': [31.8, -7.1], 'mozambique': [-25.9, 32.6], 'myanmar': [19.8, 96.2],
-  'namibia': [-22.6, 17.1], 'nauru': [-0.5, 166.9], 'navassa island': [18.4, -75.0],
-  'nepal': [28.4, 84.1], 'netherlands': [52.1, 5.3], 'new caledonia': [-22.3, 166.5],
-  'new zealand': [-41.3, 174.8], 'nicaragua': [12.1, -86.3], 'niger': [17.6, 8.1],
-  'nigeria': [9.1, 7.5], 'niue': [-19.1, -170.0], 'norfolk island': [-29.0, 167.9],
-  'north korea': [39.0, 125.8], 'north macedonia': [41.5, 21.7],
-  'northern ireland': [54.6, -5.9], 'norway': [60.5, 8.5],
-  'ogasawara': [27.1, 142.2], 'oman': [23.6, 58.5],
-  'pakistan': [30.4, 69.3], 'palau': [7.5, 134.6], 'palestine': [31.9, 35.2],
-  'palmyra & jarvis': [5.9, -162.1], 'panama': [9.0, -79.5],
-  'papua new guinea': [-6.3, 143.9], 'paraguay': [-23.4, -58.4], 'peru': [-12.0, -77.1],
-  'peter 1 island': [-68.8, -90.6], 'philippines': [14.6, 121.0],
-  'pitcairn island': [-25.1, -130.1], 'poland': [51.9, 19.1],
-  'portugal': [39.4, -8.2], 'pratas island': [20.7, 116.7],
-  'prince edward & marion': [-46.9, 37.7], 'puerto rico': [18.2, -66.6],
-  'qatar': [25.4, 51.2], 'reunion': [-21.1, 55.5], 'reunion island': [-21.1, 55.5],
-  'rodriguez island': [-19.7, 63.4], 'romania': [45.9, 25.0],
-  'rotuma island': [-12.5, 177.1], 'rwanda': [-1.9, 29.9],
-  'saba & st. eustatius': [17.6, -63.2], 'samoa': [-13.8, -172.0],
-  'san andres & providencia': [12.6, -81.7], 'san felix & san ambrosio': [-26.3, -80.1],
-  'san marino': [43.9, 12.4], 'sao tome & principe': [0.2, 6.6],
-  'sardinia': [40.1, 9.1], 'saudi arabia': [23.9, 45.1],
-  'scarborough reef': [15.1, 117.8], 'scotland': [56.5, -4.2],
-  'senegal': [14.7, -17.5], 'serbia': [44.0, 20.9], 'seychelles': [-4.7, 55.5],
-  'sierra leone': [8.5, -11.8], 'singapore': [1.4, 103.8],
-  'sint maarten': [18.0, -63.1], 'slovakia': [48.7, 19.7], 'slovenia': [46.2, 14.8],
-  'solomon islands': [-9.4, 160.0], 'somalia': [2.0, 45.3],
-  'south africa': [-30.6, 22.9], 'south cook islands': [-21.2, -159.8],
-  'south georgia island': [-54.3, -36.5], 'south korea': [35.9, 127.8],
-  'south orkney islands': [-60.6, -45.6], 'south sandwich islands': [-56.3, -26.4],
-  'south shetland islands': [-62.1, -58.7], 'south sudan': [4.9, 31.6],
-  'spain': [40.5, -3.7], 'spratly islands': [10.0, 114.0], 'sri lanka': [7.9, 80.8],
-  'st. barthelemy': [17.9, -62.8], 'st. helena': [-16.0, -5.7],
-  'st. kitts & nevis': [17.4, -62.7], 'st. lucia': [14.0, -61.0],
-  'st. martin': [18.1, -63.1], 'st. paul island': [57.2, -170.3],
-  'st. peter & st. paul': [1.0, -29.3], 'st. pierre & miquelon': [46.8, -56.2],
-  'st. vincent': [13.3, -61.2], 'sudan': [15.5, 32.6], 'suriname': [4.0, -56.0],
-  'svalbard': [77.6, 15.6], 'sweden': [60.1, 18.6], 'switzerland': [46.8, 8.2],
-  'syria': [35.0, 38.0], 'taiwan': [23.7, 120.9], 'tajikistan': [38.6, 68.8],
-  'tanzania': [-6.4, 34.9], 'temotu province': [-10.7, 166.0], 'thailand': [15.9, 100.5],
-  'togo': [6.1, 1.2], 'tokelau islands': [-9.2, -172.0], 'tonga': [-21.2, -175.2],
-  'trinidad & tobago': [10.5, -61.3], 'trindade & martim vaz': [-20.5, -29.3],
-  'tristan da cunha & gough': [-37.1, -12.3], 'tromelin island': [-15.9, 54.5],
-  'tunisia': [33.9, 9.5], 'turkey': [39.9, 32.9], 'turkmenistan': [38.0, 58.4],
-  'turks & caicos islands': [21.8, -71.8], 'tuvalu': [-7.5, 178.0],
-  'uganda': [1.4, 32.3], 'uk base areas on cyprus': [34.6, 32.9],
-  'ukraine': [48.4, 31.2], 'united arab emirates': [23.4, 53.8],
-  'united nations hq': [40.7, -74.0], 'united states': [38.9, -77.0],
-  'uruguay': [-34.9, -56.2], 'us virgin islands': [18.3, -64.9],
-  'uzbekistan': [41.3, 69.3], 'vanuatu': [-17.7, 168.3], 'vatican city': [41.9, 12.5],
-  'venezuela': [10.5, -66.9], 'vietnam': [14.1, 108.3], 'wake island': [19.3, 166.6],
-  'wales': [52.1, -3.8], 'wallis & futuna': [-13.3, -176.2],
-  'west kiribati': [1.4, 173.0], 'west malaysia': [3.1, 101.7],
-  'western sahara': [24.2, -13.0], 'willis island': [-16.3, 150.0],
-  'yemen': [15.4, 44.2], 'zambia': [-15.4, 28.3], 'zanzibar': [-6.2, 39.2],
-  'zimbabwe': [-17.8, 31.0],
-};
-
-function parseDxpeditionItem(item) {
-  // Title format: "Entity: Dates -- Callsign -- QSL via: Method"
-  const title = (item.title || '').trim();
-  const parts = title.split(/\s*--\s*/);
-  let entity = '', dateStr = '', callsign = '', qsl = '';
-
-  if (parts.length >= 1) {
-    // First segment: "Entity: Dates" or just "Entity"
-    const colonIdx = parts[0].indexOf(':');
-    if (colonIdx > 0) {
-      entity = parts[0].substring(0, colonIdx).trim();
-      dateStr = parts[0].substring(colonIdx + 1).trim();
-    } else {
-      entity = parts[0].trim();
-    }
-  }
-  if (parts.length >= 2) callsign = parts[1].trim();
-  if (parts.length >= 3) {
-    qsl = parts[2].replace(/^QSL\s*via:\s*/i, '').trim();
-  }
-
-  // Parse date range with best-effort regex
-  // Patterns: "Jan 1-Feb 16, 2026", "January 1 - February 16, 2026"
-  let startDate = null, endDate = null;
-  const now = new Date();
-  const year = now.getUTCFullYear();
-
-  const rangeMatch = dateStr.match(/(\w+)\s+(\d+)\s*[-–]\s*(\w+)\s+(\d+)(?:\s*,?\s*(\d{4}))?/);
-  if (rangeMatch) {
-    const yr = rangeMatch[5] ? parseInt(rangeMatch[5]) : year;
-    startDate = parseDatePart(rangeMatch[1], parseInt(rangeMatch[2]), yr);
-    endDate = parseDatePart(rangeMatch[3], parseInt(rangeMatch[4]), yr);
-    if (endDate) endDate.setUTCHours(23, 59, 59);
-  } else {
-    // Single month range: "Jan 1-16, 2026"
-    const singleMatch = dateStr.match(/(\w+)\s+(\d+)\s*[-–]\s*(\d+)(?:\s*,?\s*(\d{4}))?/);
-    if (singleMatch) {
-      const yr = singleMatch[4] ? parseInt(singleMatch[4]) : year;
-      startDate = parseDatePart(singleMatch[1], parseInt(singleMatch[2]), yr);
-      endDate = parseDatePart(singleMatch[1], parseInt(singleMatch[3]), yr);
-      if (endDate) endDate.setUTCHours(23, 59, 59);
-    }
-  }
-
-  const active = startDate && endDate && now >= startDate && now <= endDate;
-
-  // Look up approximate coordinates from DXCC entity name
-  const coords = DXCC_COORDS[entity.toLowerCase().trim()] || null;
-
-  return {
-    callsign,
-    entity,
-    dateStr,
-    startDate: startDate ? startDate.toISOString() : null,
-    endDate: endDate ? endDate.toISOString() : null,
-    qsl,
-    link: (item.link || '').trim(),
-    active: !!active,
-    lat: coords ? coords[0] : null,
-    lon: coords ? coords[1] : null,
-  };
-}
-
-function parseDatePart(monthStr, day, year) {
-  const months = { jan: 0, feb: 1, mar: 2, apr: 3, may: 4, jun: 5, jul: 6, aug: 7, sep: 8, oct: 9, nov: 10, dec: 11,
-    january: 0, february: 1, march: 2, april: 3, june: 5, july: 6, august: 7, september: 8, october: 9, november: 10, december: 11 };
-  const m = months[monthStr.toLowerCase()];
-  if (m === undefined || isNaN(day)) return null;
-  return new Date(Date.UTC(year, m, day));
-}
-
-app.get('/api/dxpeditions', async (req, res) => {
-  try {
-    if (dxpeditionCache.data && Date.now() < dxpeditionCache.expires) {
-      return res.json(dxpeditionCache.data);
-    }
-
-    const xml = await fetchText('https://www.ng3k.com/adxo.xml');
-    const parser = new XMLParser({ ignoreAttributes: false });
-    const parsed = parser.parse(xml);
-
-    const channel = parsed.rss && parsed.rss.channel;
-    let items = channel && channel.item;
-    if (!items) items = [];
-    if (!Array.isArray(items)) items = [items];
-
-    const results = items
-      .map(parseDxpeditionItem)
-      .filter(d => d.callsign) // skip items without a callsign
-      .sort((a, b) => {
-        // Active first, then by start date ascending
-        if (a.active !== b.active) return a.active ? -1 : 1;
-        const aDate = a.startDate ? new Date(a.startDate).getTime() : 0;
-        const bDate = b.startDate ? new Date(b.startDate).getTime() : 0;
-        return aDate - bDate;
-      });
-
-    dxpeditionCache.data = results;
-    dxpeditionCache.expires = Date.now() + DXPEDITION_TTL;
-    res.json(results);
-  } catch (err) {
-    console.error('Error fetching DXpeditions:', err.message);
-    res.status(502).json({ error: 'Failed to fetch DXpedition data' });
-  }
-});
-
-// --- Contests (WA7BNM Contest Calendar) ---
-
-const contestCache = { data: null, expires: 0 };
-const CONTEST_TTL = 6 * 60 * 60 * 1000; // 6 hours
-
-function parseContestItem(item) {
-  const name = (item.title || '').trim();
-  const description = (item.description || '').trim();
-  const link = (item.link || '').trim();
-
-  // Parse dates from description: "0000Z, Feb 7 to 2400Z, Feb 8"
-  let startDate = null, endDate = null, dateStr = description;
-  const now = new Date();
-  const year = now.getUTCFullYear();
-
-  const dateMatch = description.match(/(\d{4})Z\s*,?\s*(\w+)\s+(\d+)\s+to\s+(\d{4})Z\s*,?\s*(\w+)\s+(\d+)/i);
-  if (dateMatch) {
-    const startHour = parseInt(dateMatch[1].substring(0, 2));
-    const startMin = parseInt(dateMatch[1].substring(2));
-    const endHourRaw = parseInt(dateMatch[4].substring(0, 2));
-    const endMin = parseInt(dateMatch[4].substring(2));
-
-    startDate = parseDatePart(dateMatch[2], parseInt(dateMatch[3]), year);
-    endDate = parseDatePart(dateMatch[5], parseInt(dateMatch[6]), year);
-
-    if (startDate) {
-      startDate.setUTCHours(startHour, startMin);
-    }
-    if (endDate) {
-      // 2400Z means midnight of the next day
-      if (endHourRaw === 24) {
-        endDate.setUTCDate(endDate.getUTCDate() + 1);
-        endDate.setUTCHours(0, 0);
-      } else {
-        endDate.setUTCHours(endHourRaw, endMin);
-      }
-    }
-  }
-
-  // Infer mode from contest name
-  const upper = name.toUpperCase();
-  let mode = 'mixed';
-  if (/\bCW\b/.test(upper)) mode = 'cw';
-  else if (/\b(SSB|PHONE)\b/.test(upper)) mode = 'phone';
-  else if (/\b(RTTY|FT[48]|DIGITAL|PSK|DATA)\b/.test(upper)) mode = 'digital';
-
-  // Determine status
-  let status = 'upcoming';
-  if (startDate && endDate) {
-    if (now >= startDate && now <= endDate) status = 'active';
-    else if (now > endDate) status = 'past';
-  }
-
-  return {
-    name,
-    dateStr,
-    startDate: startDate ? startDate.toISOString() : null,
-    endDate: endDate ? endDate.toISOString() : null,
-    link,
-    mode,
-    status,
-  };
-}
-
-app.get('/api/contests', async (req, res) => {
-  try {
-    if (contestCache.data && Date.now() < contestCache.expires) {
-      return res.json(contestCache.data);
-    }
-
-    const xml = await fetchText('https://www.contestcalendar.com/contestcal.php?rss');
-    const parser = new XMLParser({ ignoreAttributes: false });
-    const parsed = parser.parse(xml);
-
-    const channel = parsed.rss && parsed.rss.channel;
-    let items = channel && channel.item;
-    if (!items) items = [];
-    if (!Array.isArray(items)) items = [items];
-
-    const results = items
-      .map(parseContestItem)
-      .filter(c => c.status !== 'past') // exclude past contests
-      .sort((a, b) => {
-        // Active first, then upcoming by start date
-        if (a.status !== b.status) {
-          if (a.status === 'active') return -1;
-          if (b.status === 'active') return 1;
-        }
-        const aDate = a.startDate ? new Date(a.startDate).getTime() : 0;
-        const bDate = b.startDate ? new Date(b.startDate).getTime() : 0;
-        return aDate - bDate;
-      });
-
-    contestCache.data = results;
-    contestCache.expires = Date.now() + CONTEST_TTL;
-    res.json(results);
-  } catch (err) {
-    console.error('Error fetching contests:', err.message);
-    res.status(502).json({ error: 'Failed to fetch contest data' });
-  }
-});
-
-// Proxy prop.kc2g.com propagation GeoJSON contours
-app.get('/api/propagation', async (req, res) => {
-  try {
-    const validTypes = ['mufd', 'fof2'];
-    const type = validTypes.includes(req.query.type) ? req.query.type : 'mufd';
-    const data = await fetchJSON(`https://prop.kc2g.com/renders/current/${type}-normal-now.geojson`);
-    res.json(data);
-  } catch (err) {
-    console.error('Error fetching propagation data:', err.message);
-    res.status(502).json({ error: 'Failed to fetch propagation data' });
-  }
-});
-
-// Proxy prop.kc2g.com MUF/foF2 bare image overlay (4096×2048 Plate Carree)
-app.get('/api/propagation/image', async (req, res) => {
-  try {
-    const validTypes = ['mufd', 'fof2'];
-    const type = validTypes.includes(req.query.type) ? req.query.type : 'mufd';
-    const url = `https://prop.kc2g.com/renders/current/${type}-bare-now.jpg`;
-    const parsed = new URL(url);
-    const resolvedIP = await resolveHost(parsed.hostname);
-    if (isPrivateIP(resolvedIP)) {
-      return res.status(403).json({ error: 'Blocked' });
-    }
-    const proxyReq = https.get({
-      hostname: resolvedIP,
-      path: parsed.pathname,
-      port: 443,
-      headers: { 'User-Agent': 'HamTab/1.0', 'Host': parsed.hostname },
-      servername: parsed.hostname,
-    }, (upstream) => {
-      if (upstream.statusCode < 200 || upstream.statusCode >= 300) {
-        upstream.resume();
-        return res.status(502).json({ error: 'Failed to fetch propagation image' });
-      }
-      res.set('Content-Type', upstream.headers['content-type'] || 'image/jpeg');
-      res.set('Cache-Control', 'public, max-age=300, s-maxage=900'); // 5m browser, 15m edge
-      upstream.pipe(res);
-    });
-    proxyReq.on('error', (err) => {
-      console.error('Propagation image proxy error:', err.message);
-      if (!res.headersSent) res.status(502).json({ error: 'Failed to fetch propagation image' });
-    });
-    proxyReq.setTimeout(15000, () => { proxyReq.destroy(); });
-  } catch (err) {
-    console.error('Error fetching propagation image:', err.message);
-    res.status(502).json({ error: 'Failed to fetch propagation image' });
-  }
-});
-
-// D-RAP (D Region Absorption Prediction) image overlay — NOAA SWPC
-// Shows HF radio absorption caused by solar X-ray and proton events
-app.get('/api/drap/image', async (req, res) => {
-  try {
-    const url = 'https://services.swpc.noaa.gov/images/animations/d-rap/global/latest.png';
-    const parsed = new URL(url);
-    const resolvedIP = await resolveHost(parsed.hostname);
-    if (isPrivateIP(resolvedIP)) {
-      return res.status(403).json({ error: 'Blocked' });
-    }
-    const proxyReq = https.get({
-      hostname: resolvedIP,
-      path: parsed.pathname,
-      port: 443,
-      headers: { 'User-Agent': 'HamTab/1.0', 'Host': parsed.hostname },
-      servername: parsed.hostname,
-    }, (upstream) => {
-      if (upstream.statusCode < 200 || upstream.statusCode >= 300) {
-        upstream.resume();
-        return res.status(502).json({ error: 'Failed to fetch D-RAP image' });
-      }
-      res.set('Content-Type', upstream.headers['content-type'] || 'image/png');
-      res.set('Cache-Control', 'public, max-age=300, s-maxage=900'); // 5m browser, 15m edge
-      upstream.pipe(res);
-    });
-    proxyReq.on('error', (err) => {
-      console.error('D-RAP image proxy error:', err.message);
-      if (!res.headersSent) res.status(502).json({ error: 'Failed to fetch D-RAP image' });
-    });
-    proxyReq.setTimeout(15000, () => { proxyReq.destroy(); });
-  } catch (err) {
-    console.error('Error fetching D-RAP image:', err.message);
-    res.status(502).json({ error: 'Failed to fetch D-RAP image' });
-  }
-});
-
-// D-RAP text grid — parsed frequency data for client-side heatmap rendering
-// 4° lat/lon grid of "Highest Frequency Affected by 1dB Absorption" in MHz
-const drapDataCache = { data: null, expires: 0 };
-const DRAP_TTL = 5 * 60 * 1000; // 5 min — SWPC updates ~every 1-2 min but data changes slowly
-
-app.get('/api/drap/data', async (req, res) => {
-  try {
-    if (drapDataCache.data && Date.now() < drapDataCache.expires) {
-      return res.json(drapDataCache.data);
-    }
-    const url = 'https://services.swpc.noaa.gov/text/drap_global_frequencies.txt';
-    const text = await secureFetch(url);
-
-    // Parse the NOAA text grid into a structured object
-    const lines = text.split('\n');
-    const lons = []; // longitude values from header
-    const rows = []; // { lat, values[] }
-
-    for (const line of lines) {
-      const trimmed = line.trim();
-      if (!trimmed || trimmed.startsWith('#') || trimmed.startsWith('---')) continue;
-
-      // Header row: longitude values
-      if (lons.length === 0 && /^\s*-?\d+/.test(trimmed) && !trimmed.includes('|')) {
-        trimmed.split(/\s+/).forEach(v => lons.push(parseFloat(v)));
-        continue;
-      }
-
-      // Data row: "lat | val val val ..."
-      const match = trimmed.match(/^\s*(-?\d+)\s*\|\s*(.+)$/);
-      if (match) {
-        const lat = parseInt(match[1]);
-        const values = match[2].trim().split(/\s+/).map(Number);
-        rows.push({ lat, values });
-      }
-    }
-
-    const result = { lons, rows };
-    drapDataCache.data = result;
-    drapDataCache.expires = Date.now() + DRAP_TTL;
-    res.json(result);
-  } catch (err) {
-    console.error('Error fetching D-RAP data:', err.message);
-    res.status(502).json({ error: 'Failed to fetch D-RAP data' });
-  }
-});
-
-// (Weather radar + cloud tiles moved to server/routes/weather.js)
-
-// --- VOACAP prediction engine ---
-
-// SSN cache — refreshed every 24 hours from NOAA predicted monthly sunspot number
-const ssnCache = { data: null, expires: 0 };
-const SSN_TTL = 24 * 60 * 60 * 1000; // 24 hours
-
-async function getCurrentSSN() {
-  if (ssnCache.data && Date.now() < ssnCache.expires) {
-    return ssnCache.data;
-  }
-
-  try {
-    // NOAA predicted monthly sunspot number (JSON array of [year, month, ssn, ...])
-    const raw = await fetchJSON('https://services.swpc.noaa.gov/json/predicted_monthly_sunspot_number.json');
-    const now = new Date();
-    const currentYear = now.getUTCFullYear();
-    const currentMonth = now.getUTCMonth() + 1;
-
-    // Find entry matching current UTC month/year
-    let entry = null;
-    if (Array.isArray(raw)) {
-      entry = raw.find(e =>
-        e['time-tag'] &&
-        new Date(e['time-tag']).getUTCFullYear() === currentYear &&
-        (new Date(e['time-tag']).getUTCMonth() + 1) === currentMonth
-      );
-    }
-
-    if (entry && entry['predicted_ssn'] != null) {
-      ssnCache.data = {
-        ssn: parseFloat(entry['predicted_ssn']),
-        month: currentMonth,
-        year: currentYear,
-        source: 'noaa',
-      };
-    } else {
-      // Fallback: use first entry or a reasonable default
-      const fallbackSSN = Array.isArray(raw) && raw.length > 0 && raw[0]['predicted_ssn'] != null
-        ? parseFloat(raw[0]['predicted_ssn'])
-        : 50;
-      ssnCache.data = {
-        ssn: fallbackSSN,
-        month: currentMonth,
-        year: currentYear,
-        source: 'noaa-fallback',
-      };
-    }
-
-    ssnCache.expires = Date.now() + SSN_TTL;
-    return ssnCache.data;
-  } catch (err) {
-    console.error('Error fetching SSN:', err.message);
-
-    // If we have stale data, return it
-    if (ssnCache.data) return ssnCache.data;
-
-    // Last resort: reasonable default
-    return {
-      ssn: 50,
-      month: new Date().getUTCMonth() + 1,
-      year: new Date().getUTCFullYear(),
-      source: 'default',
-    };
-  }
-}
-
-// --- K-index cache for real-time geomagnetic corrections ---
-// K-index from HamQSL updates every 3 hours but conditions can change rapidly,
-// so we use a 10-minute cache to balance freshness with API load.
-const kIndexCache = { kindex: null, aindex: null, expires: 0 };
-const KINDEX_TTL = 10 * 60 * 1000; // 10 minutes
-
-async function getCurrentKIndex() {
-  // Return cached data if still valid
-  if (Date.now() < kIndexCache.expires && kIndexCache.kindex !== null) {
-    return { kindex: kIndexCache.kindex, aindex: kIndexCache.aindex };
-  }
-
-  try {
-    const xml = await fetchText('https://www.hamqsl.com/solarxml.php');
-    const solar = parseSolarXML(xml);
-
-    const kindex = parseInt(solar.indices.kindex, 10);
-    const aindex = parseInt(solar.indices.aindex, 10);
-
-    kIndexCache.kindex = isNaN(kindex) ? null : kindex;
-    kIndexCache.aindex = isNaN(aindex) ? null : aindex;
-    kIndexCache.expires = Date.now() + KINDEX_TTL;
-
-    return { kindex: kIndexCache.kindex, aindex: kIndexCache.aindex };
-  } catch (err) {
-    console.error('Error fetching K-index:', err.message);
-
-    // If we have stale data, return it
-    if (kIndexCache.kindex !== null) {
-      return { kindex: kIndexCache.kindex, aindex: kIndexCache.aindex };
-    }
-
-    // Return null if no data available (caller should handle gracefully)
-    return { kindex: null, aindex: null };
-  }
-}
-
-// Calculate effective SSN with K-index degradation for geomagnetic storms.
-// During storms, absorption increases and MUF decreases. We model this
-// by reducing the "effective" SSN used for propagation calculations.
-//
-// Research basis (NOAA G-scale, SWSC aviation study):
-// - G1 (Kp=5): Minor storm, degradation begins
-// - G2 (Kp=6): "HF fading at higher latitudes"
-// - G3 (Kp=7): "HF may be intermittent" — ~50% MUF drop threshold
-// - G4 (Kp=8): "HF sporadic" — severe degradation
-// - G5 (Kp=9): "HF may be impossible" — near blackout
-// Sources: swpc.noaa.gov/noaa-scales-explanation, swsc-journal.org 2022
-function calculateEffectiveSSN(baseSSN, kIndex) {
-  // K-index to degradation factor mapping based on NOAA G-scale research:
-  // - MOD threshold: 30% MUF drop, SEV threshold: 50% MUF drop
-  // - "Storm-time decrease falls to 0.3-0.5 of undisturbed levels"
-  const degradationMap = {
-    0: 1.00,  // quiet — no effect
-    1: 1.00,  // quiet — no effect
-    2: 1.00,  // quiet — no measurable HF impact per research
-    3: 0.95,  // unsettled — minimal effect (~5%)
-    4: 0.90,  // unsettled/active — minor effect (~10%)
-    5: 0.80,  // G1 minor storm — degradation begins (~20%)
-    6: 0.65,  // G2 moderate — "fading at higher latitudes" (~35%)
-    7: 0.50,  // G3 strong — "intermittent" = ~50% MUF drop threshold
-    8: 0.30,  // G4 severe — "sporadic" (~70% loss)
-    9: 0.15,  // G5 extreme — "may be impossible" (~85% loss)
-  };
-
-  // If K-index unavailable, use base SSN unchanged
-  if (kIndex === null || kIndex === undefined || kIndex < 0 || kIndex > 9) {
-    return {
-      effectiveSSN: baseSSN,
-      baseSSN,
-      kIndex: null,
-      degradationFactor: 1.0,
-      degradationPercent: 0,
-    };
-  }
-
-  const factor = degradationMap[kIndex] || 1.0;
-  const effectiveSSN = Math.round(baseSSN * factor * 10) / 10;
-  const degradationPercent = Math.round((1 - factor) * 100);
-
-  return {
-    effectiveSSN,
-    baseSSN,
-    kIndex,
-    degradationFactor: factor,
-    degradationPercent,
-  };
-}
-
-// Classify K-index into bands for cache key partitioning:
-// q=quiet (K<=2), u=unsettled (K=3), s=storm (K=4-5), x=extreme (K>=6)
-function getKBand(kIndex) {
-  if (kIndex === null || kIndex === undefined) return 'q'; // assume quiet if unknown
-  if (kIndex <= 2) return 'q';
-  if (kIndex === 3) return 'u';
-  if (kIndex <= 5) return 's';
-  return 'x';
-}
-
-// VOACAP prediction cache — keyed by rounded params, TTL 1 hour
-const voacapCache = {}; // { key: { data, expires } }
-const VOACAP_TTL = 60 * 60 * 1000; // 1 hour
-
-
-// Simplified propagation model (server-side fallback when Python unavailable)
-// Mirrors the client-side calculate24HourMatrix from band-conditions.js
-function simplifiedVoacapMatrix(txLat, txLon, ssn, month, power, mode, toaDeg, longPath) {
-  // Mode SNR/bandwidth mapping
-  const modeParams = {
-    CW:  { snr: 39, bw: 500 },
-    SSB: { snr: 73, bw: 2700 },
-    FT8: { snr: 2,  bw: 50 },
-  };
-  const mp = modeParams[mode] || modeParams.SSB;
-
-  // Simplified solar flux estimate from SSN (empirical: SFI ≈ 63 + 0.73 * SSN)
-  const sfi = 63 + 0.73 * ssn;
-
-  // VOACAP band subset (no 160m, no 60m)
-  const bands = [
-    { name: '80m',  freq: 3.7 },
-    { name: '40m',  freq: 7.15 },
-    { name: '30m',  freq: 10.12 },
-    { name: '20m',  freq: 14.15 },
-    { name: '17m',  freq: 18.1 },
-    { name: '15m',  freq: 21.2 },
-    { name: '12m',  freq: 24.93 },
-    { name: '10m',  freq: 28.5 },
-  ];
-
-  const matrix = [];
-
-  for (let hour = 0; hour < 24; hour++) {
-    // Solar zenith–based day fraction
-    const df = dayFractionServer(txLat, txLon, hour);
-    const muf = calculateMUFServer(sfi, df);
-
-    const bandResults = {};
-    for (const band of bands) {
-      let rel = calculateBandReliabilityServer(band.freq, muf, df >= 0.5, {
-        mode, powerWatts: power, toaDeg, longPath,
-      });
-      bandResults[band.name] = { rel: Math.round(rel), snr: 0, mode: '' };
-    }
-
-    matrix.push({
-      hour,
-      bands: bandResults,
-      muf: Math.round(muf * 10) / 10,
-    });
-  }
-
-  return matrix;
-}
-
-// Server-side day fraction (mirrors client band-conditions.js dayFraction)
-function dayFractionServer(lat, lon, utcHour) {
-  if (lat == null || lon == null) {
-    return (utcHour >= 6 && utcHour < 18) ? 1.0 : 0.0;
-  }
-  const now = new Date();
-  const start = new Date(now.getFullYear(), 0, 1);
-  const dayOfYear = Math.floor((now - start) / 86400000) + 1;
-  const declRad = Math.asin(
-    Math.sin(23.44 * Math.PI / 180) * Math.sin((360 / 365) * (dayOfYear - 81) * Math.PI / 180)
-  );
-  const solarNoonOffset = lon / 15;
-  const hourAngle = (utcHour - 12 + solarNoonOffset) * 15;
-  const haRad = hourAngle * Math.PI / 180;
-  const latRad = lat * Math.PI / 180;
-  const cosZenith = Math.sin(latRad) * Math.sin(declRad) +
-                    Math.cos(latRad) * Math.cos(declRad) * Math.cos(haRad);
-  const zenith = Math.acos(Math.max(-1, Math.min(1, cosZenith))) * 180 / Math.PI;
-  if (zenith <= 80) return 1.0;
-  if (zenith >= 100) return 0.0;
-  return (100 - zenith) / 20;
-}
-
-// Server-side MUF calculation (mirrors client band-conditions.js)
-function calculateMUFServer(sfi, dayFrac) {
-  if (dayFrac === true) dayFrac = 1.0;
-  else if (dayFrac === false) dayFrac = 0.0;
-  const foF2Factor = 0.6 + 0.3 * dayFrac;
-  const foF2 = foF2Factor * Math.sqrt(Math.max(sfi, 50));
-  return foF2 * 3.5; // obliquity factor
-}
-
-// Server-side band reliability (mirrors client band-conditions.js)
-// Tuned to be more realistic — original was too optimistic (all green).
-// Key changes:
-// - Above MUF: much steeper dropoff (was 40% base, now 15%)
-// - FT8 bonus: reduced from +30 to +15 (still significant but not magic)
-// - Below optimal: reduced daytime floor, increased night bonus modestly
-function calculateBandReliabilityServer(freqMHz, muf, isDay, opts) {
-  const mufLower = muf * 0.5;   // LUF approximation
-  const mufOptimal = muf * 0.85; // optimal frequency ~85% of MUF
-  let base = 0;
-
-  if (freqMHz < mufLower) {
-    // Below LUF: heavy D-layer absorption during day, better at night
-    if (isDay) {
-      base = Math.max(0, 10 - (mufLower - freqMHz) * 3); // was 20, now 10 max
-    } else {
-      base = Math.min(70, 40 + (mufLower - freqMHz) * 1.0); // reduced from 85/60
-    }
-  } else if (freqMHz <= mufOptimal) {
-    // Between LUF and optimal: good propagation zone
-    const position = (freqMHz - mufLower) / (mufOptimal - mufLower);
-    base = 50 + (35 * Math.sin(position * Math.PI)); // was 70+30, now 50+35
-  } else if (freqMHz <= muf) {
-    // Between optimal and MUF: still usable but degrading
-    const position = (freqMHz - mufOptimal) / (muf - mufOptimal);
-    base = 75 - (position * 35); // was 90-40, now 75-35 (ends at 40% at MUF)
-  } else {
-    // Above MUF: rapid dropoff — signals won't reflect
-    const excess = freqMHz - muf;
-    base = Math.max(0, 15 - excess * 5); // was 40-3x, now 15-5x (much steeper)
-  }
-
-  if (opts) {
-    // Mode adjustments — FT8 helps but isn't magic
-    if (opts.mode === 'CW') base += 8;      // was +10
-    else if (opts.mode === 'FT8') base += 15; // was +30 (way too high)
-
-    // Power adjustment (±10 dB = ±15% reliability)
-    if (opts.powerWatts && opts.powerWatts !== 100) {
-      base += 10 * Math.log10(opts.powerWatts / 100) * 1.5;
-    }
-
-    // TOA adjustment — higher angles slightly better for skip
-    if (opts.toaDeg != null) {
-      base += (opts.toaDeg - 5) * 0.5; // reduced from 1.5
-    }
-
-    // Long path penalty
-    if (opts.longPath) base -= 30; // was -25
-  }
-
-  return Math.max(0, Math.min(100, base));
-}
-
-// Representative global targets for multi-target VOACAP overview.
-// 4 targets × 24 hours = 96 predictions — fits within container CPU budget.
-const VOACAP_TARGETS = [
-  { name: 'Europe',        lat: 50.0,  lon: 10.0 },
-  { name: 'East Asia',     lat: 35.0,  lon: 135.0 },
-  { name: 'South America', lat: -15.0, lon: -47.0 },
-  { name: 'North America', lat: 40.0,  lon: -100.0 },
-];
-
-// Return SSN data
-app.get('/api/voacap/ssn', async (req, res) => {
-  try {
-    const ssn = await getCurrentSSN();
-    res.json(ssn);
-  } catch (err) {
-    console.error('Error fetching SSN:', err.message);
-    res.status(502).json({ error: 'Failed to fetch SSN data' });
-  }
-});
-
-// VOACAP bridge diagnostics — helps debug container startup issues
-app.get('/api/voacap/status', (req, res) => {
-  res.json(voacap.getStatus());
-});
-
-// VOACAP prediction endpoint
-app.get('/api/voacap', async (req, res) => {
-  try {
-    // Validate params
-    const txLat = parseFloat(req.query.txLat);
-    const txLon = parseFloat(req.query.txLon);
-    if (isNaN(txLat) || isNaN(txLon) || txLat < -90 || txLat > 90 || txLon < -180 || txLon > 180) {
-      return res.status(400).json({ error: 'Invalid or missing txLat/txLon' });
-    }
-
-    const rxLat = req.query.rxLat != null ? parseFloat(req.query.rxLat) : null;
-    const rxLon = req.query.rxLon != null ? parseFloat(req.query.rxLon) : null;
-    if (rxLat != null && (isNaN(rxLat) || rxLat < -90 || rxLat > 90)) {
-      return res.status(400).json({ error: 'Invalid rxLat' });
-    }
-    if (rxLon != null && (isNaN(rxLon) || rxLon < -180 || rxLon > 180)) {
-      return res.status(400).json({ error: 'Invalid rxLon' });
-    }
-
-    const validPowers = [5, 100, 1000];
-    const power = validPowers.includes(parseInt(req.query.power)) ? parseInt(req.query.power) : 100;
-
-    const validModes = ['CW', 'SSB', 'FT8'];
-    const mode = validModes.includes(req.query.mode) ? req.query.mode : 'SSB';
-
-    const validToa = [3, 5, 10, 15];
-    const toa = validToa.includes(parseInt(req.query.toa)) ? parseInt(req.query.toa) : 5;
-
-    const validPath = ['SP', 'LP'];
-    const pathType = validPath.includes(req.query.path) ? req.query.path : 'SP';
-    const longPath = pathType === 'LP';
-
-    // Optional client-provided SNR override (integer dB, clamped to safe range)
-    const clientSnr = req.query.snr != null ? parseInt(req.query.snr, 10) : null;
-    const snrOverride = (clientSnr != null && !isNaN(clientSnr)) ? Math.max(-10, Math.min(100, clientSnr)) : null;
-
-    // API token validation — if VOACAP_API_TOKENS is set in .env, require a valid token.
-    // Format: comma-separated list of tokens, e.g. VOACAP_API_TOKENS=abc123,def456
-    // Client sends token via X-Voacap-Token header or ?token= query param.
-    // When not configured (lanmode default), all requests are allowed.
-    const allowedTokens = process.env.VOACAP_API_TOKENS;
-    if (allowedTokens) {
-      const tokenSet = new Set(allowedTokens.split(',').map(t => t.trim()).filter(Boolean));
-      const clientToken = req.headers['x-voacap-token'] || req.query.token || '';
-      if (!tokenSet.has(clientToken)) {
-        return res.status(401).json({ error: 'Invalid or missing API token' });
-      }
-    }
-
-    // Fetch SSN and K-index in parallel for real-time corrections
-    const [ssnData, kIndexData] = await Promise.all([
-      getCurrentSSN(),
-      getCurrentKIndex(),
-    ]);
-
-    const baseSSN = ssnData.ssn;
-    const month = ssnData.month;
-    const kIndex = kIndexData.kindex;
-
-    // Calculate effective SSN with K-index degradation
-    const ssnCorrection = calculateEffectiveSSN(baseSSN, kIndex);
-    const ssn = ssnCorrection.effectiveSSN;
-    const kBand = getKBand(kIndex);
-
-    // Cache key — round lat/lon to 1 decimal, include K-band for storm transitions
-    const cacheKey = [
-      Math.round(txLat * 10) / 10,
-      Math.round(txLon * 10) / 10,
-      rxLat != null ? Math.round(rxLat * 10) / 10 : 'all',
-      rxLon != null ? Math.round(rxLon * 10) / 10 : 'all',
-      power, mode, toa, pathType, kBand,
-      snrOverride != null ? snrOverride : 'default',
-    ].join(':');
-
-    const cached = voacapCache[cacheKey];
-    if (cached && Date.now() < cached.expires) {
-      return res.json(cached.data);
-    }
-
-    // Mode → default required SNR (dB above noise in BW) and bandwidth.
-    // These are the "Normal" (level 3) defaults. Client can override via ?snr= param.
-    // See SENSITIVITY_LEVELS in voacap.js for the full 1-5 preset table.
-    const modeMap = {
-      CW:  { snr: 30, bw: 500 },
-      SSB: { snr: 54, bw: 2700 },
-      FT8: { snr: 2,  bw: 50 },
-    };
-    const modeDefaults = modeMap[mode] || modeMap.SSB;
-    const requiredSnr = snrOverride != null ? snrOverride : modeDefaults.snr;
-    const bandwidthHz = modeDefaults.bw;
-
-    // VOACAP band frequencies (no 160m, no 60m)
-    const frequencies = [3.7, 7.15, 10.12, 14.15, 18.1, 21.2, 24.93, 28.5];
-    const bandNames = ['80m', '40m', '30m', '20m', '17m', '15m', '12m', '10m'];
-
-    // Determine targets
-    let targets;
-    if (rxLat != null && rxLon != null) {
-      targets = [{ name: 'target', lat: rxLat, lon: rxLon }];
-    } else {
-      // Use representative global targets; swap NA for Caribbean if TX is in NA
-      targets = VOACAP_TARGETS.map(t => {
-        if (t.name === 'North America' && txLat > 24 && txLat < 50 && txLon > -130 && txLon < -60) {
-          return { name: 'Caribbean', lat: 18.0, lon: -66.0 };
-        }
-        return t;
-      });
-    }
-
-    // Hybrid approach: try dvoacap with a short inline timeout.
-    // On lanmode (fast CPU), dvoacap completes in seconds — return it inline.
-    // On hostedmode (slow container), timeout fires — return simplified immediately,
-    // let dvoacap continue in the background and cache for the next request.
-    const INLINE_TIMEOUT_MS = 15000; // 15 seconds — enough for lanmode, fast-fail for container
-
-    let engine = 'simplified';
-    let matrix;
-    let fallbackReason = null;
-
-    const predictionParams = {
-      tx_lat: txLat,
-      tx_lon: txLon,
-      targets: targets.map(t => ({ name: t.name, lat: t.lat, lon: t.lon })),
-      ssn,
-      month,
-      power,
-      min_angle_deg: toa,
-      long_path: longPath,
-      required_snr: requiredSnr,
-      bandwidth_hz: bandwidthHz,
-      frequencies,
-      band_names: bandNames,
-    };
-
-    if (voacap.isAvailable()) {
-      // Start the prediction — we'll race it against a timeout
-      const predictionPromise = voacap.predictMatrix(predictionParams);
-
-      try {
-        const result = await Promise.race([
-          predictionPromise,
-          new Promise((_, reject) =>
-            setTimeout(() => reject(new Error('inline timeout')), INLINE_TIMEOUT_MS)
-          ),
-        ]);
-
-        // dvoacap completed in time — use it
-        if (result.ok && result.matrix) {
-          engine = 'dvoacap';
-          matrix = result.matrix;
-        } else {
-          fallbackReason = `predictMatrix returned: ${result.error || 'no matrix'}`;
-          matrix = simplifiedVoacapMatrix(txLat, txLon, ssn, month, power, mode, toa, longPath);
-        }
-      } catch (err) {
-        // Inline timeout or prediction error — return simplified now
-        matrix = simplifiedVoacapMatrix(txLat, txLon, ssn, month, power, mode, toa, longPath);
-
-        if (err.message === 'inline timeout') {
-          fallbackReason = 'dvoacap too slow — running in background';
-          // Let the prediction continue in the background — cache result when done
-          predictionPromise.then(result => {
-            if (result.ok && result.matrix) {
-              console.log('[VOACAP] Background prediction complete — caching dvoacap result');
-              const bgResponse = {
-                engine: 'dvoacap',
-                ssn: Math.round(ssn * 10) / 10,
-                ssnBase: Math.round(baseSSN * 10) / 10,
-                kIndex: ssnCorrection.kIndex,
-                kDegradation: ssnCorrection.degradationPercent,
-                month,
-                matrix: result.matrix,
-              };
-              voacapCache[cacheKey] = { data: bgResponse, expires: Date.now() + VOACAP_TTL };
-            }
-          }).catch(bgErr => {
-            console.error(`[VOACAP] Background prediction error: ${bgErr.message}`);
-            if (bgErr.message === 'Request timed out') {
-              voacap.killAndRespawn('Background prediction timed out after 180s');
-            }
-          });
-        } else {
-          fallbackReason = `predictMatrix threw: ${err.message}`;
-        }
-      }
-    } else {
-      fallbackReason = 'bridge not available';
-      matrix = simplifiedVoacapMatrix(txLat, txLon, ssn, month, power, mode, toa, longPath);
-    }
-
-    const response = {
-      engine,
-      ssn: Math.round(ssn * 10) / 10,
-      ssnBase: Math.round(baseSSN * 10) / 10,
-      kIndex: ssnCorrection.kIndex,
-      kDegradation: ssnCorrection.degradationPercent,
-      month,
-      matrix,
-    };
-    if (fallbackReason) response.fallbackReason = fallbackReason;
-
-    const ttl = engine === 'dvoacap' ? VOACAP_TTL : 30 * 1000; // 1h vs 30s
-    voacapCache[cacheKey] = { data: response, expires: Date.now() + ttl };
-
-    res.json(response);
-  } catch (err) {
-    console.error('Error in /api/voacap:', err.message);
-    res.status(500).json({ error: 'Failed to compute VOACAP predictions' });
-  }
-});
-
-// --- Lunar math (simplified Meeus algorithms) ---
-
-// Lunar position via Jean Meeus, "Astronomical Algorithms" 2nd ed.
-// Chapters 47 (position) & 48 (illumination). Coefficients are the
-// principal terms from Table 47.A; lower-order terms omitted for speed.
-
-// Greenwich Mean Sidereal Time in degrees from Julian Date
-// Meeus, Chapter 12
-function gmst(JD) {
-  const T = (JD - 2451545.0) / 36525.0;
-  return mod360(280.46061837 + 360.98564736629 * (JD - 2451545.0)
-    + 0.000387933 * T * T - T * T * T / 38710000);
-}
-
-// Compute moon equatorial coordinates (RA, Dec) and horizontal parallax
-// at an arbitrary Date. Extracted from computeLunar for reuse in rise/set.
-function moonEquatorial(date) {
-  const JD = julianDate(date);
-  const T = (JD - 2451545.0) / 36525.0;
-  const rad = Math.PI / 180;
-
-  const Lp = mod360(218.3165 + 481267.8813 * T);
-  const Mp = mod360(134.9634 + 477198.8676 * T);
-  const D = mod360(297.8502 + 445267.1115 * T);
-  const M = mod360(357.5291 + 35999.0503 * T);
-  const F = mod360(93.2720 + 483202.0175 * T);
-
-  let lambda = Lp
-    + 6.289 * Math.sin(Mp * rad)
-    - 1.274 * Math.sin((2 * D - Mp) * rad)
-    + 0.658 * Math.sin(2 * D * rad)
-    - 0.214 * Math.sin(2 * Mp * rad)
-    - 0.186 * Math.sin(M * rad)
-    - 0.114 * Math.sin(2 * F * rad);
-  lambda = mod360(lambda);
-
-  const beta = 5.128 * Math.sin(F * rad)
-    + 0.281 * Math.sin((Mp + F) * rad)
-    - 0.277 * Math.sin((Mp - F) * rad)
-    - 0.173 * Math.sin((2 * D - F) * rad);
-
-  const horizParallax = 0.9508
-    + 0.0518 * Math.cos(Mp * rad)
-    + 0.0095 * Math.cos((2 * D - Mp) * rad)
-    + 0.0078 * Math.cos(2 * D * rad)
-    + 0.0028 * Math.cos(2 * Mp * rad);
-
-  const epsilon = 23.4393 - 0.0130 * T;
-  const lambdaRad = lambda * rad;
-  const betaRad = beta * rad;
-  const epsilonRad = epsilon * rad;
-
-  const sinDec = Math.sin(betaRad) * Math.cos(epsilonRad)
-    + Math.cos(betaRad) * Math.sin(epsilonRad) * Math.sin(lambdaRad);
-  const declination = Math.asin(sinDec) / rad;
-
-  const sinRA = Math.cos(betaRad) * Math.sin(lambdaRad) * Math.cos(epsilonRad)
-    - Math.sin(betaRad) * Math.sin(epsilonRad);
-  const cosRA = Math.cos(betaRad) * Math.cos(lambdaRad);
-  const rightAscension = mod360(Math.atan2(sinRA, cosRA) / rad);
-
-  // Sun's ecliptic longitude (approximate) — needed for phase/illumination
-  const sunLongitude = mod360(280.4665 + 36000.7698 * T);
-
-  return { rightAscension, declination, horizParallax, lambda, beta, JD, T,
-    sunLongitude, Mp, D, M, F, Lp };
-}
-
-// Compute moon horizontal coordinates (azimuth, elevation) from observer
-// Meeus, Chapter 13 — sidereal time + coordinate transformation
-function moonHorizontal(date, lat, lon) {
-  const eq = moonEquatorial(date);
-  const rad = Math.PI / 180;
-
-  const lst = mod360(gmst(eq.JD) + lon); // local sidereal time (degrees)
-  const ha = mod360(lst - eq.rightAscension); // hour angle (degrees)
-
-  const haRad = ha * rad;
-  const decRad = eq.declination * rad;
-  const latRad = lat * rad;
-
-  const sinAlt = Math.sin(latRad) * Math.sin(decRad)
-    + Math.cos(latRad) * Math.cos(decRad) * Math.cos(haRad);
-  const altitude = Math.asin(sinAlt) / rad;
-
-  // Azimuth measured from North, clockwise
-  const cosAlt = Math.cos(altitude * rad);
-  let azimuth = 0;
-  if (Math.abs(cosAlt) > 1e-10) {
-    const sinAz = -Math.cos(decRad) * Math.sin(haRad) / cosAlt;
-    const cosAz = (Math.sin(decRad) - Math.sin(latRad) * sinAlt) / (Math.cos(latRad) * cosAlt);
-    azimuth = mod360(Math.atan2(sinAz, cosAz) / rad);
-  }
-
-  return { azimuth, elevation: altitude, horizParallax: eq.horizParallax };
-}
-
-// Compute next moonrise and moonset from current time
-// Searches forward 48 hours in 5-minute steps, then linear-interpolates
-// Standard altitude: h0 = 0.7275 * parallax - 34/60 (refraction + semidiameter)
-function computeMoonRiseSet(lat, lon) {
-  const now = new Date();
-  const STEP_MS = 5 * 60000; // 5 minutes
-  const STEPS = 576; // 48 hours / 5 min = 576 steps
-
-  let prevCorr = null;
-  let rise = null, set = null;
-
-  for (let i = 0; i <= STEPS && (!rise || !set); i++) {
-    const t = new Date(now.getTime() + i * STEP_MS);
-    const horiz = moonHorizontal(t, lat, lon);
-    const h0 = 0.7275 * horiz.horizParallax - 34 / 60; // degrees
-    const corr = horiz.elevation - h0;
-
-    if (prevCorr !== null) {
-      if (prevCorr < 0 && corr >= 0 && !rise) {
-        // Moon rising — interpolate
-        const frac = -prevCorr / (corr - prevCorr);
-        rise = Math.floor((now.getTime() + ((i - 1) + frac) * STEP_MS) / 1000);
-      }
-      if (prevCorr >= 0 && corr < 0 && !set) {
-        // Moon setting — interpolate
-        const frac = prevCorr / (prevCorr - corr);
-        set = Math.floor((now.getTime() + ((i - 1) + frac) * STEP_MS) / 1000);
-      }
-    }
-    prevCorr = corr;
-  }
-
-  return { moonrise: rise, moonset: set };
-}
-
-function computeLunar(lat, lon) {
-  const now = new Date();
-  const eq = moonEquatorial(now);
-  const rad = Math.PI / 180;
-
-  const distance = 6378.14 / Math.sin(eq.horizParallax * rad); // km
-
-  let elongation = eq.lambda - eq.sunLongitude;
-  elongation = mod360(elongation + 180) - 180;
-
-  const illumination = (1 - Math.cos(elongation * rad)) / 2 * 100;
-  const phase = moonPhaseName(elongation);
-
-  const avgDistance = 384400; // km — mean Earth-Moon distance
-  const pathLoss = 20 * Math.log10(distance / avgDistance);
-
-  const result = {
-    phase,
-    illumination: Math.round(illumination * 10) / 10,
-    declination: Math.round(eq.declination * 10) / 10,
-    distance: Math.round(distance),
-    pathLoss: Math.round(pathLoss * 100) / 100,
-    elongation: Math.round(elongation * 10) / 10,
-    eclipticLon: Math.round(eq.lambda * 10) / 10,
-    eclipticLat: Math.round(eq.beta * 100) / 100,
-    rightAscension: Math.round(eq.rightAscension * 10) / 10,
-  };
-
-  // Observer-dependent fields — only computed when lat/lon provided
-  if (lat !== undefined && lon !== undefined) {
-    const horiz = moonHorizontal(now, lat, lon);
-    result.azimuth = Math.round(horiz.azimuth * 10) / 10;
-    result.elevation = Math.round(horiz.elevation * 10) / 10;
-
-    const riseSet = computeMoonRiseSet(lat, lon);
-    result.moonrise = riseSet.moonrise; // Unix timestamp or null
-    result.moonset = riseSet.moonset;   // Unix timestamp or null
-  }
-
-  return result;
-}
-
-function julianDate(date) {
-  const Y = date.getUTCFullYear();
-  const M = date.getUTCMonth() + 1;
-  const D = date.getUTCDate()
-    + date.getUTCHours() / 24
-    + date.getUTCMinutes() / 1440
-    + date.getUTCSeconds() / 86400;
-
-  let y = Y, m = M;
-  if (m <= 2) { y -= 1; m += 12; }
-  const A = Math.floor(y / 100);
-  const B = 2 - A + Math.floor(A / 4);
-  return Math.floor(365.25 * (y + 4716)) + Math.floor(30.6001 * (m + 1)) + D + B - 1524.5;
-}
-
-function mod360(x) {
-  return ((x % 360) + 360) % 360;
-}
-
-function moonPhaseName(elongation) {
-  const e = mod360(elongation);
-  if (e < 22.5) return 'New Moon';
-  if (e < 67.5) return 'Waxing Crescent';
-  if (e < 112.5) return 'First Quarter';
-  if (e < 157.5) return 'Waxing Gibbous';
-  if (e < 202.5) return 'Full Moon';
-  if (e < 247.5) return 'Waning Gibbous';
-  if (e < 292.5) return 'Last Quarter';
-  if (e < 337.5) return 'Waning Crescent';
-  return 'New Moon';
-}
-
-// (Hardened fetch helpers moved to server/services/http-fetch.js)
-
-// --- SOTA summit coordinate cache (TTL-based) ---
-
-const sotaSummitCache = {};  // { 'W7W/LC-001': { lat, lon, expires }, ... }
-const SUMMIT_TTL_OK   = 24 * 60 * 60 * 1000; // 24 hours for successful lookups
-const SUMMIT_TTL_FAIL =      60 * 60 * 1000; // 1 hour for failed lookups
-
-async function fetchSummitCoords(associationCode, summitCode) {
-  const key = `${associationCode}/${summitCode}`;
-  const cached = sotaSummitCache[key];
-  if (cached && Date.now() < cached.expires) return cached;
-  try {
-    const data = await fetchJSON(
-      `https://api2.sota.org.uk/api/summits/${encodeURIComponent(associationCode)}/${encodeURIComponent(summitCode)}`
-    );
-    const entry = {
-      lat: data.latitude ?? null,
-      lon: data.longitude ?? null,
-      expires: Date.now() + SUMMIT_TTL_OK,
-    };
-    sotaSummitCache[key] = entry;
-    return entry;
-  } catch {
-    const entry = { lat: null, lon: null, expires: Date.now() + SUMMIT_TTL_FAIL };
-    sotaSummitCache[key] = entry;
-    return entry;
-  }
-}
-
-function parseSolarXML(xml) {
-  const parser = new XMLParser({
-    ignoreAttributes: false,
-    attributeNamePrefix: '@_',
-    textNodeName: '#text',
-  });
-
-  const doc = parser.parse(xml);
-  const sd = doc.solar.solardata;
-
-  const str = (v) => (v != null ? String(v).trim() : ''); // normalize XML text nodes to trimmed strings (or '' if missing)
-
-  const indices = {
-    sfi:           str(sd.solarflux),
-    sunspots:      str(sd.sunspots),
-    aindex:        str(sd.aindex),
-    kindex:        str(sd.kindex),
-    xray:          str(sd.xray),
-    signalnoise:   str(sd.signalnoise),
-    updated:       str(sd.updated),
-    solarwind:     str(sd.solarwind),
-    magneticfield: str(sd.magneticfield),
-    protonflux:    str(sd.protonflux),
-    electonflux:   str(sd.electonflux),
-    aurora:        str(sd.aurora),
-    latdegree:     str(sd.latdegree),
-    heliumline:    str(sd.heliumline),
-    geomagfield:   str(sd.geomagfield),
-    kindexnt:      str(sd.kindexnt),
-    fof2:          str(sd.fof2),
-    muffactor:     str(sd.muffactor),
-    muf:           str(sd.muf),
-  };
-
-  // Parse HF band conditions
-  const bands = [];
-  const rawBands = sd.calculatedconditions?.band;
-  if (rawBands) {
-    const bandArr = Array.isArray(rawBands) ? rawBands : [rawBands];
-    bandArr.forEach(b => {
-      bands.push({
-        band:      b['@_name']  || '',
-        time:      b['@_time']  || '',
-        condition: String(typeof b === 'object' ? b['#text'] : b || '').trim(),
-      });
-    });
-  }
-
-  // Parse VHF conditions
-  const vhf = [];
-  const rawVhf = sd.calculatedvhfconditions?.phenomenon;
-  if (rawVhf) {
-    const vhfArr = Array.isArray(rawVhf) ? rawVhf : [rawVhf];
-    vhfArr.forEach(p => {
-      vhf.push({
-        name:      p['@_name']     || '',
-        location:  p['@_location'] || '',
-        condition: String(typeof p === 'object' ? p['#text'] : p || '').trim(),
-      });
-    });
-  }
-
-  return { indices, bands, vhf };
-}
-
-// --- Cache eviction (register all caches with the shared eviction service) ---
-[dxcCallCache, pskHeardCache, satPassCache, satTleCache, sdoDirCache, sotaSummitCache, voacapCache, dxpeditionCache, contestCache].forEach(registerCache);
+// --- Cache eviction ---
 startEviction();
 
 // --- Server startup ---
